Data.Row.Extras

Description

Instances for Rec and Var types

Documentation

newtype JsonRec s

Constructors

JsonRec

Fields

unJsonRec :: Rec s

newtype JsonVar s

Constructors

JsonVar

Fields

unJsonVar :: Var s

namedBranchFromJSON :: forall s. (AllUniqueLabels s, Forall s FromJSON) => String -> Value -> Parser (Var s)

Parse a 'Var s' from JSON if the label of the branch is known.

type family (l :: Row k) .\\ (r :: Row k) :: Row k where ... infixl 6

Fast diff. The implementation in row-types is exponential in time and memory in the number of overlapping rows, due to limitations in ghc's handling of type families. This version is much faster.

Type level Row difference. That is, l .\\ r is the row remaining after removing any matching elements of r from l.

Equations

('R l) .\\ ('R r) = 'R (Diff l r)