Plutus.Contract

Documentation

newtype Contract w (s :: Row *) e a

Contract w s e a is a contract with schema s, producing a value of type a or an error e. See note [Contract Schema].

Constructors

Contract

Fields

unContract :: Eff (ContractEffs w e) a

data ContractError

Constructors

WalletContractError WalletAPIError

ChainIndexContractError Text ChainIndexResponse

ConstraintResolutionContractError MkTxError

ToCardanoConvertContractError ToCardanoError

ResumableContractError MatchingError

CCheckpointContractError CheckpointError

EndpointDecodeContractError

Fields

eeEndpointDescription :: EndpointDescription

The endpoint description which the decoding error occurred from

eeEndpointValue :: EndpointValue Value

The endpoint value that was used as an endpoint parameter

eeErrorMessage :: Text

JSON decoding error message

OtherContractError Text

class AsContractError r where

Minimal complete definition

_ContractError

Methods

_ContractError :: Prism' r ContractError

_WalletContractError :: Prism' r WalletAPIError

_ChainIndexContractError :: Prism' r (Text, ChainIndexResponse)

_ConstraintResolutionContractError :: Prism' r MkTxError

_ToCardanoConvertContractError :: Prism' r ToCardanoError

_ResumableContractError :: Prism' r MatchingError

_CCheckpointContractError :: Prism' r CheckpointError

_EndpointDecodeContractError :: Prism' r (EndpointDescription, EndpointValue Value, Text)

_OtherContractError :: Prism' r Text

class IsContract c where

Class of types that can be trivially converted to a Contract. For use with functions where it is convenient to accept both Contract and Promise types.

Methods

toContract :: c w s e a -> Contract w s e a

(>>) :: Monad m => m a -> m b -> m b infixl 1

Sequentially compose two actions, discarding any value produced by the first, like sequencing operators (such as the semicolon) in imperative languages.

'as >> bs' can be understood as the do expression

do as
   bs

throwError :: MonadError e m => e -> m a

handleError :: forall w s e e' a. (e -> Contract w s e' a) -> Contract w s e a -> Contract w s e' a

Handle errors, potentially throwing new errors.

mapError :: forall w s e e' a. (e -> e') -> Contract w s e a -> Contract w s e' a

Transform any exceptions thrown by the Contract using the given function.

runError :: forall w s e e0 a. Contract w s e a -> Contract w s e0 (Either e a)

Turn a contract with error type e and return type a into one with any error type (ie. throwing no errors) that returns 'Either e a'

Select

data Promise w (s :: Row *) e a

A wrapper indicating that this contract starts with a waiting action. For use with select.

awaitPromise :: Promise w s e a -> Contract w s e a

promiseMap :: (Contract w1 s1 e1 a1 -> Contract w2 s2 e2 a2) -> Promise w1 s1 e1 a1 -> Promise w2 s2 e2 a2

Lift a mapping function for Contract to a mapping function for Promise.

promiseBind :: Promise w s e a -> (a -> Contract w s e b) -> Promise w s e b

Run more Contract code after the Promise.

both :: Promise w s e a -> Promise w s e b -> Promise w s e (a, b)

Execute both contracts in any order

selectEither :: forall w s e a b. Promise w s e a -> Promise w s e b -> Promise w s e (Either a b)

A variant of select for contracts with different return types.

select :: forall w s e a. Promise w s e a -> Promise w s e a -> Promise w s e a

select returns the contract that makes progress first, discarding the other one.

However, note that if multiples promises are chained together like P1 select P2 select P3 and all three can make progress at the same moment, then select will prioritize the promises starting from the right (first P3 then P2 then P1).

selectList :: [Promise w s e a] -> Contract w s e a

selectList returns the contract that makes progress first, discarding the other ones.

However, if multiple contracts can make progress, selectList prioritizes the ones appearing first in the input list. Therefore, the order of the list of promises is important.

never :: Promise w s e a

A Promise that is never fulfilled. This is the identity of select.

Dealing with time

awaitSlot :: forall w s e. AsContractError e => Slot -> Contract w s e Slot

Wait until the slot

isSlot :: forall w s e. AsContractError e => Slot -> Promise w s e Slot

Wait until the slot

currentSlot :: forall w s e. AsContractError e => Contract w s e Slot

Deprecated: Use currentNodeClientSlot instead

Get the current slot number

currentPABSlot :: forall w s e. AsContractError e => Contract w s e Slot

Deprecated: Use currentNodeClientSlot instead

Get the current slot number of PAB

currentNodeClientSlot :: forall w s e. AsContractError e => Contract w s e Slot

Get the current slot number of the node client (the local or remote node) that the application is connected to.

currentChainIndexSlot :: forall w s e. AsContractError e => Contract w s e Slot

Get the current node slot number querying slot number from plutus chain index to be aligned with slot at local running node

waitNSlots :: forall w s e. AsContractError e => Natural -> Contract w s e Slot

Wait for a number of slots to pass

awaitTime :: forall w s e. AsContractError e => POSIXTime -> Contract w s e POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

Example: if starting time is 0 and slot length is 3s, then `awaitTime 4` waits until slot 2 and returns the value `POSIXTime 5`.

isTime :: forall w s e. AsContractError e => POSIXTime -> Promise w s e POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

currentTime :: forall w s e. AsContractError e => Contract w s e POSIXTime

Deprecated: Use currentNodeClientTimeRange instead

Get the latest time of the current slot.

Example: if slot length is 3s and current slot is 2, then currentTime returns the value `POSIXTime 5`

currentNodeClientTimeRange :: forall w s e. AsContractError e => Contract w s e (POSIXTime, POSIXTime)

Get the POSIXTime range of the current slot.

Example: if slot length is 3s and current slot is 2, then currentTimeRange returns the time interval [3, 5[.

waitNMilliSeconds :: forall w s e. AsContractError e => DiffMilliSeconds -> Contract w s e POSIXTime

Wait for a number of milliseconds starting at the ending time of the current slot, and return the latest time we know has passed.

Example: if starting time is 0, slot length is 3000ms and current slot is 0, then `waitNMilliSeconds 0` returns the value `POSIXTime 2000` and `waitNMilliSeconds 1000` returns the value `POSIXTime 5`.

Endpoints

type HasEndpoint l a s = (HasType l (EndpointValue a) (Input s), HasType l ActiveEndpoint (Output s), KnownSymbol l, ContractRow s)

newtype EndpointDescription

Constructors

EndpointDescription

Fields

getEndpointDescription :: String

type Endpoint l a = l .== (EndpointValue a, ActiveEndpoint)

endpoint :: forall l a w s e b. (HasEndpoint l a s, AsContractError e, FromJSON a) => (a -> Contract w s e b) -> Promise w s e b

Expose an endpoint, return the data that was entered

handleEndpoint :: forall l a w s e1 e2 b. (HasEndpoint l a s, AsContractError e1, FromJSON a) => (Either e1 a -> Contract w s e2 b) -> Promise w s e2 b

endpointWithMeta :: forall l a w s e meta b. (HasEndpoint l a s, AsContractError e, ToJSON meta, FromJSON a) => meta -> (a -> Contract w s e b) -> Promise w s e b

Expose an endpoint with some metadata. Return the data that was entered.

type EmptySchema = Empty

Blockchain events

watchAddressUntilSlot :: forall w s e. AsContractError e => CardanoAddress -> Slot -> Contract w s e (Map TxOutRef DecoratedTxOut)

Wait until the target slot and get the unspent transaction outputs at an address.

watchAddressUntilTime :: forall w s e. AsContractError e => CardanoAddress -> POSIXTime -> Contract w s e (Map TxOutRef DecoratedTxOut)

Wait until the target time and get the unspent transaction outputs at an address.

fundsAtAddressGt :: forall w s e. AsContractError e => CardanoAddress -> Value -> Contract w s e (Map TxOutRef DecoratedTxOut)

Watch an address for changes, and return the outputs at that address when the total value at the address has surpassed the given value.

fundsAtAddressGeq :: forall w s e. AsContractError e => CardanoAddress -> Value -> Contract w s e (Map TxOutRef DecoratedTxOut)

Watch an address for changes, and return the outputs at that address when the total value at the address has reached or surpassed the given value.

awaitUtxoSpent :: forall w s e. AsContractError e => TxOutRef -> Contract w s e ChainIndexTx

Wait until the UTXO has been spent, returning the transaction that spends it.

utxoIsSpent :: forall w s e. AsContractError e => TxOutRef -> Promise w s e ChainIndexTx

Wait until the UTXO has been spent, returning the transaction that spends it.

awaitUtxoProduced :: forall w s e. AsContractError e => CardanoAddress -> Contract w s e (NonEmpty ChainIndexTx)

Wait until one or more unspent outputs are produced at an address.

utxoIsProduced :: forall w s e. AsContractError e => CardanoAddress -> Promise w s e (NonEmpty ChainIndexTx)

Wait until one or more unspent outputs are produced at an address.

Chain index requests

datumFromHash :: forall w s e. AsContractError e => DatumHash -> Contract w s e (Maybe Datum)

datumsAt :: forall w s e. AsContractError e => CardanoAddress -> Contract w s e [Datum]

Get the all datums at an address whether or not the corresponding utxo have been consumed or not.

findReferenceValidatorScripByHash :: forall w s e. AsContractError e => ValidatorHash -> CardanoAddress -> Contract w s e TxOutRef

Find the reference to an utxo containing a reference script by its the script hash, amongst the utxos at a given address

validatorFromHash :: forall w s e. AsContractError e => ValidatorHash -> Contract w s e (Maybe (Versioned Validator))

mintingPolicyFromHash :: forall w s e. AsContractError e => MintingPolicyHash -> Contract w s e (Maybe (Versioned MintingPolicy))

stakeValidatorFromHash :: forall w s e. AsContractError e => StakeValidatorHash -> Contract w s e (Maybe (Versioned StakeValidator))

txOutFromRef :: forall w s e. AsContractError e => TxOutRef -> Contract w s e (Maybe DecoratedTxOut)

unspentTxOutFromRef :: forall w s e. AsContractError e => TxOutRef -> Contract w s e (Maybe DecoratedTxOut)

txFromTxId :: forall w s e. AsContractError e => TxId -> Contract w s e (Maybe ChainIndexTx)

utxoRefMembership :: forall w s e. AsContractError e => TxOutRef -> Contract w s e IsUtxoResponse

utxoRefsAt :: forall w s e. AsContractError e => PageQuery TxOutRef -> CardanoAddress -> Contract w s e UtxosResponse

Get the unspent transaction output references at an address.

utxoRefsWithCurrency :: forall w s e. AsContractError e => PageQuery TxOutRef -> AssetClass -> Contract w s e UtxosResponse

Get the unspent transaction output references with a specific currrency (AssetClass).

utxosAt :: forall w s e. AsContractError e => CardanoAddress -> Contract w s e (Map TxOutRef DecoratedTxOut)

Get the unspent transaction outputs at an address.

utxosTxOutTxFromTx :: AsContractError e => ChainIndexTx -> Contract w s e [(TxOutRef, (DecoratedTxOut, ChainIndexTx))]

Get the unspent transaction outputs from a ChainIndexTx.

getTip :: forall w s e. AsContractError e => Contract w s e Tip

Wallet's information

ownPaymentPubKeyHash :: forall w s e. AsContractError e => Contract w s e PaymentPubKeyHash

Deprecated: Use ownFirstPaymentPubKeyHash, ownPaymentPubKeyHashes or ownAddresses instead

Get the hash of a public key belonging to the wallet that runs this contract. * Any funds paid to this public key hash will be treated as the wallet's own funds * The wallet is able to sign transactions with the private key of this public key, for example, if the public key is added to the requiredSignatures field of Tx. * There is a 1-n relationship between wallets and public keys (although in the mockchain n=1)

ownPaymentPubKeyHashes :: forall w s e. AsContractError e => Contract w s e [PaymentPubKeyHash]

ownFirstPaymentPubKeyHash :: forall w s e. AsContractError e => Contract w s e PaymentPubKeyHash

ownAddresses :: forall w s e. AsContractError e => Contract w s e (NonEmpty CardanoAddress)

Get the addresses belonging to the wallet that runs this contract. * Any funds paid to one of these addresses will be treated as the wallet's own funds * The wallet is able to sign transactions with the private key of one of its public key, for example, if the public key is added to the requiredSignatures field of Tx. * There is a 1-n relationship between wallets and addresses (although in the mockchain n=1)

ownAddress :: forall w s e. AsContractError e => Contract w s e CardanoAddress

Get the first address of the wallet that runs this contract.

ownUtxos :: forall w s e. AsContractError e => Contract w s e (Map TxOutRef DecoratedTxOut)

Get all utxos belonging to the wallet that runs this contract.

Contract instance Id

data ContractInstanceId

Unique ID for contract instance

ownInstanceId :: forall w s e. AsContractError e => Contract w s e ContractInstanceId

Get the ContractInstanceId of this instance.

Notifications

tell :: w -> Contract w s e ()

Update the contract's accumulating state w

Transactions

data WalletAPIError

An error thrown by wallet interactions.

adjustUnbalancedTx :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e UnbalancedTx

Adjust the unbalanced tx

submitTx :: forall w s e. AsContractError e => TxConstraints Void Void -> Contract w s e CardanoTx

Build a transaction that satisfies the constraints, then submit it to the network. The constraints do not refer to any typed script inputs or outputs.

submitTxConfirmed :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e ()

A version of submitTx that waits until the transaction has been confirmed on the ledger before returning.

submitTxConstraints :: forall a w s e. (ToData (RedeemerType a), FromData (DatumType a), ToData (DatumType a), AsContractError e) => TypedValidator a -> TxConstraints (RedeemerType a) (DatumType a) -> Contract w s e CardanoTx

Build a transaction that satisfies the constraints, then submit it to the network. Using the current outputs at the contract address and the contract's own public key to solve the constraints.

submitTxConstraintsSpending :: forall a w s e. (ToData (RedeemerType a), FromData (DatumType a), ToData (DatumType a), AsContractError e) => TypedValidator a -> Map TxOutRef DecoratedTxOut -> TxConstraints (RedeemerType a) (DatumType a) -> Contract w s e CardanoTx

Build a transaction that satisfies the constraints using the UTXO map to resolve any input constraints (see InputConstraint)

submitTxConstraintsWith :: forall a w s e. (ToData (RedeemerType a), FromData (DatumType a), ToData (DatumType a), AsContractError e) => ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> Contract w s e CardanoTx

Build a transaction that satisfies the constraints, then submit it to the network. Using the given constraints.

submitUnbalancedTx :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e CardanoTx

Send an unbalanced transaction to be balanced and signed. Returns the ID of the final transaction when the transaction was submitted. Throws an error if balancing or signing failed.

submitBalancedTx :: forall w s e. AsContractError e => CardanoTx -> Contract w s e CardanoTx

Send an balanced transaction to be signed. Returns the ID of the final transaction when the transaction was submitted. Throws an error if signing failed.

balanceTx :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e CardanoTx

Send an unbalanced transaction to be balanced. Returns the balanced transaction. Throws an error if balancing failed.

mkTxConstraints :: forall a w s e. (ToData (RedeemerType a), FromData (DatumType a), ToData (DatumType a), AsContractError e) => ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> Contract w s e UnbalancedTx

Build a transaction that satisfies the constraints

yieldUnbalancedTx :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e ()

Take an UnbalancedTx then balance, sign and submit it to the blockchain without returning any results.

Tx confirmation

awaitTxConfirmed :: forall w s e. AsContractError e => TxId -> Contract w s e ()

Wait until a transaction is confirmed (added to the ledger). If the transaction is never added to the ledger then awaitTxConfirmed never returns

awaitTxStatusChange :: forall w s e. AsContractError e => TxId -> Contract w s e TxStatus

Wait for the status of a transaction to change

isTxConfirmed :: forall w s e. AsContractError e => TxId -> Promise w s e ()

Wait until a transaction is confirmed (added to the ledger).

Tx output confirmation

awaitTxOutStatusChange :: forall w s e. AsContractError e => TxOutRef -> Contract w s e TxOutStatus

Wait for the status of a transaction output to change.

Parameters

getParams :: forall w s e. AsContractError e => Contract w s e Params

Get the configured parameter set.

Checkpoints

checkpoint :: forall w s e a. (AsCheckpointError e, FromJSON a, ToJSON a) => Contract w s e a -> Contract w s e a

Write the current state of the contract to a checkpoint.

checkpointLoop :: forall w s e a b. (AsCheckpointError e, FromJSON a, ToJSON a, ToJSON b, FromJSON b) => (a -> Contract w s e (Either b a)) -> a -> Contract w s e b

class AsCheckpointError r where

Methods

_CheckpointError :: Prism' r CheckpointError

data CheckpointError

Constructors

JSONDecodeError Text

Logging

module Plutus.Contract.Logging

Row-related things

class (r .! l) ≈ a => HasType (l :: Symbol) (a :: k) (r :: Row k)

type ContractRow s = (AllUniqueLabels (Input s), AllUniqueLabels (Output s))

Constraints on the contract schema, ensuring that the labels of the schema are unique.

type family (l :: Row k) .\/ (r :: Row k) :: Row k where ...

Equations

(x :: Row k) .\/ ('R ('[] :: [LT k]) :: Row k) = x

('R ('[] :: [LT k]) :: Row k) .\/ (y :: Row k) = y

('R l :: Row a) .\/ ('R r :: Row a) = 'R (MinJoinR l r)

type Empty = 'R ('[] :: [LT a])