Plutus.Contract.Checkpoint

Checkpoints

This module contains a checkpoints mechanism that can be used to store intermediate results of Eff programs as JSON values inside a CheckpointStore. It works similar to the short-circuiting behavior of Error: Before we execute an action Eff effs a whose result should be checkpointed, we check if the there is already a value of a for this checkpoint it in the store. If there is, we return it instead of running the action. If there isn't, we run the action a and then store the result.

To create a checkpoint use jsonCheckpoint.

To handle the checkpoint effect use handleCheckpoint.

data Checkpoint r where

Constructors

DoCheckpoint :: Checkpoint ()

AllocateKey :: Checkpoint CheckpointKey

Store :: ToJSON a => CheckpointKey -> CheckpointKey -> a -> Checkpoint ()

Retrieve :: FromJSON a => CheckpointKey -> Checkpoint (Either CheckpointError (Maybe a))

data CheckpointError

Constructors

JSONDecodeError Text

class AsCheckpointError r where

Minimal complete definition

_CheckpointError

Methods

_CheckpointError :: Prism' r CheckpointError

_JSONDecodeError :: Prism' r Text

newtype CheckpointStore

Constructors

CheckpointStore

Fields

unCheckpointStore :: Map CheckpointKey (CheckpointStoreItem Value)

data CheckpointStoreItem a

Constructors

CheckpointStoreItem

Fields

csValue :: a

csNewKey :: CheckpointKey

data CheckpointKey

data CheckpointLogMsg

Constructors

LogFoundValueRestoringKey CheckpointKey

LogDecodingErrorAtKey CheckpointKey

LogNoValueForKey CheckpointKey

LogDoCheckpoint

LogAllocateKey

LogRetrieve CheckpointKey

LogStore CheckpointKey CheckpointKey

LogKeyUpdate CheckpointKey CheckpointKey

jsonCheckpoint

Arguments

:: forall err a effs. (Member Checkpoint effs, Member (Error err) effs, ToJSON a, FromJSON a, AsCheckpointError err)

=> Eff effs a
The action that is checkpointed

-> Eff effs a

Create a checkpoint for an action. handleCheckpoint (jsonCheckpoint action) will

Obtain a CheckpointKey that identifies the position of the current checkpoint in the program

Run action, convert its result to JSON and store it in the checkpoint store if there is no value at the key

Retrieve the result as a JSON value from the store, parse it, and return it *instead* of running action if there is a value at the key.

jsonCheckpointLoop

Arguments

:: forall err a b effs. (Member Checkpoint effs, Member (Error err) effs, ToJSON a, FromJSON a, ToJSON b, FromJSON b, AsCheckpointError err)

=> (a -> Eff effs (Either b a))
The action that is repeated until it returns a Left. Only the accumulated result of the action will be stored.

-> a
Initial value

-> Eff effs b

handleCheckpoint :: forall effs. (Member (State CheckpointStore) effs, Member (State CheckpointKey) effs, Member (LogMsg CheckpointLogMsg) effs) => Eff (Checkpoint ': effs) ~> Eff effs

Handle the Checkpoint effect in terms of CheckpointStore and CheckpointKey states.

completedIntervals :: CheckpointStore -> IntervalSet (Interval CheckpointKey)

Intervals of checkpoint keys that are completely covered by the checkpoint store.

maxKey :: CheckpointStore -> Maybe CheckpointKey

The maximum key that is present in the store