Plutus.Contract.Effects

Plutus application backend request effect types

data PABReq

Requests that Contracts can make

Constructors

AdjustUnbalancedTxReq UnbalancedTx

AwaitSlotReq Slot

AwaitTimeReq POSIXTime

AwaitUtxoSpentReq TxIn

AwaitUtxoProducedReq CardanoAddress

AwaitTxStatusChangeReq TxId

AwaitTxOutStatusChangeReq TxIn

CurrentNodeClientSlotReq

CurrentChainIndexSlotReq

CurrentTimeReq

CurrentNodeClientTimeRangeReq

GetParamsReq

OwnContractInstanceIdReq

OwnAddressesReq

ChainIndexQueryReq ChainIndexQuery

BalanceTxReq UnbalancedTx

WriteBalancedTxReq CardanoTx

ExposeEndpointReq ActiveEndpoint

PosixTimeRangeToContainedSlotRangeReq POSIXTimeRange

YieldUnbalancedTxReq UnbalancedTx

_AdjustUnbalancedTxReq :: Prism' PABReq UnbalancedTx

_AwaitSlotReq :: Prism' PABReq Slot

_AwaitTimeReq :: Prism' PABReq POSIXTime

_AwaitUtxoSpentReq :: Prism' PABReq TxIn

_AwaitUtxoProducedReq :: Prism' PABReq CardanoAddress

_CurrentNodeClientSlotReq :: Prism' PABReq ()

_CurrentChainIndexSlotReq :: Prism' PABReq ()

_CurrentTimeReq :: Prism' PABReq ()

_CurrentNodeClientTimeRangeReq :: Prism' PABReq ()

_AwaitTxStatusChangeReq :: Prism' PABReq TxId

_AwaitTxOutStatusChangeReq :: Prism' PABReq TxIn

_GetParamsReq :: Prism' PABReq ()

_OwnContractInstanceIdReq :: Prism' PABReq ()

_OwnAddressesReq :: Prism' PABReq ()

_ChainIndexQueryReq :: Prism' PABReq ChainIndexQuery

_BalanceTxReq :: Prism' PABReq UnbalancedTx

_WriteBalancedTxReq :: Prism' PABReq CardanoTx

_ExposeEndpointReq :: Prism' PABReq ActiveEndpoint

_PosixTimeRangeToContainedSlotRangeReq :: Prism' PABReq POSIXTimeRange

_YieldUnbalancedTxReq :: Prism' PABReq UnbalancedTx

Chain index query effect types

_DatumFromHash :: Prism' ChainIndexQuery DatumHash

_ValidatorFromHash :: Prism' ChainIndexQuery ValidatorHash

_MintingPolicyFromHash :: Prism' ChainIndexQuery MintingPolicyHash

_RedeemerFromHash :: Prism' ChainIndexQuery RedeemerHash

_TxOutFromRef :: Prism' ChainIndexQuery TxOutRef

_TxFromTxId :: Prism' ChainIndexQuery TxId

_UnspentTxOutFromRef :: Prism' ChainIndexQuery TxOutRef

_UtxoSetMembership :: Prism' ChainIndexQuery TxOutRef

_UtxoSetAtAddress :: Prism' ChainIndexQuery (PageQuery TxOutRef, CardanoAddress)

_UtxoSetWithCurrency :: Prism' ChainIndexQuery (PageQuery TxOutRef, AssetClass)

_TxsFromTxIds :: Prism' ChainIndexQuery [TxId]

_TxoSetAtAddress :: Prism' ChainIndexQuery (PageQuery TxOutRef, CardanoAddress)

_GetTip :: Prism' ChainIndexQuery ()

Plutus application backend response effect types

data PABResp

Responses that Contracts receive

Constructors

AdjustUnbalancedTxResp UnbalancedTx

AwaitSlotResp Slot

AwaitTimeResp POSIXTime

AwaitUtxoSpentResp ChainIndexTx

AwaitUtxoProducedResp (NonEmpty ChainIndexTx)

AwaitTxStatusChangeResp TxId TxStatus

AwaitTxOutStatusChangeResp TxIn TxOutStatus

CurrentNodeClientSlotResp Slot

CurrentChainIndexSlotResp Slot

CurrentTimeResp POSIXTime

CurrentNodeClientTimeRangeResp (POSIXTime, POSIXTime)

GetParamsResp Params

OwnContractInstanceIdResp ContractInstanceId

OwnAddressesResp (NonEmpty CardanoAddress)

ChainIndexQueryResp ChainIndexResponse

BalanceTxResp BalanceTxResponse

WriteBalancedTxResp WriteBalancedTxResponse

ExposeEndpointResp EndpointDescription (EndpointValue Value)

PosixTimeRangeToContainedSlotRangeResp (Either SlotConversionError SlotRange)

YieldUnbalancedTxResp ()

_AdjustUnbalancedTxResp :: Prism' PABResp UnbalancedTx

_AwaitSlotResp :: Prism' PABResp Slot

_AwaitTimeResp :: Prism' PABResp POSIXTime

_AwaitUtxoSpentResp :: Prism' PABResp ChainIndexTx

_AwaitUtxoProducedResp :: Prism' PABResp (NonEmpty ChainIndexTx)

_CurrentNodeClientSlotResp :: Prism' PABResp Slot

_CurrentChainIndexSlotResp :: Prism' PABResp Slot

_CurrentTimeResp :: Prism' PABResp POSIXTime

_CurrentNodeClientTimeRangeResp :: Prism' PABResp (POSIXTime, POSIXTime)

_AwaitTxStatusChangeResp :: Prism' PABResp (TxId, TxStatus)

_AwaitTxStatusChangeResp' :: TxId -> Prism' PABResp TxStatus

_AwaitTxOutStatusChangeResp :: Prism' PABResp (TxIn, TxOutStatus)

_GetParamsResp :: Prism' PABResp Params

_OwnContractInstanceIdResp :: Prism' PABResp ContractInstanceId

_OwnAddressesResp :: Prism' PABResp (NonEmpty CardanoAddress)

_ChainIndexQueryResp :: Prism' PABResp ChainIndexResponse

_BalanceTxResp :: Prism' PABResp BalanceTxResponse

_WriteBalancedTxResp :: Prism' PABResp WriteBalancedTxResponse

_ExposeEndpointResp :: Prism' PABResp (EndpointDescription, EndpointValue Value)

_PosixTimeRangeToContainedSlotRangeResp :: Prism' PABResp (Either SlotConversionError SlotRange)

_YieldUnbalancedTxResp :: Prism' PABResp ()

Chain index response effect types

_DatumHashResponse :: Prism' ChainIndexResponse (Maybe Datum)

_ValidatorHashResponse :: Prism' ChainIndexResponse (Maybe (Versioned Validator))

_MintingPolicyHashResponse :: Prism' ChainIndexResponse (Maybe (Versioned MintingPolicy))

_RedeemerHashResponse :: Prism' ChainIndexResponse (Maybe Redeemer)

_TxOutRefResponse :: Prism' ChainIndexResponse (Maybe DecoratedTxOut)

_UnspentTxOutResponse :: Prism' ChainIndexResponse (Maybe DecoratedTxOut)

_TxIdResponse :: Prism' ChainIndexResponse (Maybe ChainIndexTx)

_UtxoSetMembershipResponse :: Prism' ChainIndexResponse IsUtxoResponse

_UtxoSetAtResponse :: Prism' ChainIndexResponse UtxosResponse

_UtxoSetWithCurrencyResponse :: Prism' ChainIndexResponse UtxosResponse

_TxIdsResponse :: Prism' ChainIndexResponse [ChainIndexTx]

_TxoSetAtResponse :: Prism' ChainIndexResponse TxosResponse

_GetTipResponse :: Prism' ChainIndexResponse Tip

Etc.

matches :: PABReq -> PABResp -> Bool

data ChainIndexQuery

Represents all possible chain index queries. Each constructor contains the input(s) needed for the query. These possible queries correspond to the constructors of the data type ChainIndexQueryEffect.

Constructors

DatumFromHash DatumHash

ValidatorFromHash ValidatorHash

MintingPolicyFromHash MintingPolicyHash

StakeValidatorFromHash StakeValidatorHash

RedeemerFromHash RedeemerHash

TxOutFromRef TxOutRef

UnspentTxOutFromRef TxOutRef

TxFromTxId TxId

UtxoSetMembership TxOutRef

UtxoSetAtAddress (PageQuery TxOutRef) CardanoAddress

UnspentTxOutSetAtAddress (PageQuery TxOutRef) CardanoAddress

DatumsAtAddress (PageQuery TxOutRef) CardanoAddress

UtxoSetWithCurrency (PageQuery TxOutRef) AssetClass

TxsFromTxIds [TxId]

TxoSetAtAddress (PageQuery TxOutRef) CardanoAddress

GetTip

data ChainIndexResponse

Represents all possible responses to chain index queries. Each constructor contain the output resulting for the chain index query. These possible responses come from the data type ChainIndexQueryEffect.

Constructors

DatumHashResponse (Maybe Datum)

ValidatorHashResponse (Maybe (Versioned Validator))

MintingPolicyHashResponse (Maybe (Versioned MintingPolicy))

StakeValidatorHashResponse (Maybe (Versioned StakeValidator))

TxOutRefResponse (Maybe DecoratedTxOut)

UnspentTxOutResponse (Maybe DecoratedTxOut)

RedeemerHashResponse (Maybe Redeemer)

TxIdResponse (Maybe ChainIndexTx)

UtxoSetMembershipResponse IsUtxoResponse

UtxoSetAtResponse UtxosResponse

UnspentTxOutsAtResponse (QueryResponse [(TxOutRef, DecoratedTxOut)])

DatumsAtResponse (QueryResponse [Datum])

UtxoSetWithCurrencyResponse UtxosResponse

TxIdsResponse [ChainIndexTx]

TxoSetAtResponse TxosResponse

GetTipResponse Tip

data BalanceTxResponse

Constructors

BalanceTxFailed WalletAPIError

BalanceTxSuccess CardanoTx

balanceTxResponse :: Iso' BalanceTxResponse (Either WalletAPIError CardanoTx)

data WriteBalancedTxResponse

Constructors

WriteBalancedTxFailed WalletAPIError

WriteBalancedTxSuccess CardanoTx

writeBalancedTxResponse :: Iso' WriteBalancedTxResponse (Either WalletAPIError CardanoTx)

data ActiveEndpoint

Constructors

ActiveEndpoint

Fields

aeDescription :: EndpointDescription

The name of the endpoint

aeMetadata :: Maybe Value

Data that should be shown to the user