Plutus.Contract.Resumable

The Resumable effect

data Resumable i o r where

A data type for representing non-deterministic prompts.

Constructors

RRequest :: o -> Resumable i o i

RSelect :: Resumable i o Bool

RZero :: Resumable i o a

prompt :: Member (Resumable i o) effs => o -> Eff effs i

select :: forall i o effs a. Member (Resumable i o) effs => Eff effs a -> Eff effs a -> Eff effs a

never :: forall i o effs a. Member (Resumable i o) effs => Eff effs a

Handling the Resumable effect

data Request o

Constructors

Request

Fields

rqID :: RequestID

itID :: IterationID

rqRequest :: o

data Response i

Constructors

Response

Fields

rspRqID :: RequestID

rspItID :: IterationID

rspResponse :: i

newtype RequestID

A value that uniquely identifies requests made during the execution of Resumable programs.

Constructors

RequestID Natural

newtype IterationID

A value that uniquely identifies groups of requests.

Constructors

IterationID Natural

newtype Requests o

Constructors

Requests

Fields

unRequests :: [Request o]

type ResumableEffs i o effs a = NonDet ': (State IterationID ': (State RequestID ': (State (ReqMap i o effs a) ': (State (Requests o) ': effs))))

newtype Responses i

Constructors

Responses

Fields

unResponses :: Map (IterationID, RequestID) i

insertResponse :: Response i -> Responses i -> Responses i

responses :: Responses i -> [Response i]

A list of all responses ordered by iteration and request ID

_Responses :: forall i. Iso' (Responses i) (Map (IterationID, RequestID) i)

Handling the Resumable effect with continuations

handleResumable :: forall i o effs. (Member (Yield o i) effs, Member NonDet effs) => Eff (Resumable i o ': effs) ~> Eff effs

Interpret the Resumable effect in terms of the Yield and NonDet effects.

suspendNonDet :: forall i o a effs. Eff (Yield o i ': ResumableEffs i o effs a) a -> Eff effs (Maybe (MultiRequestContStatus i o effs a))

Interpret Yield as a prompt-type effect using NonDet to branch out and choose a branch, and the State effects to keep track of request IDs.

data MultiRequestContStatus i o effs a

Status of a suspended MultiRequestContinuation.

Constructors

AResult a
Done

AContinuation (MultiRequestContinuation i o effs a)
Waiting for inputs

data MultiRequestContinuation i o effs a

A continuation that accepts a response to one of several requests.

Constructors

MultiRequestContinuation

Fields

ndcCont :: Response i -> Eff effs (Maybe (MultiRequestContStatus i o effs a))

Continuation for the response

ndcRequests :: Requests o

The list of all open requests.

newtype ReqMap i o effs a

A map of requests to continuations. For each request, identified by a pair of RequestID and IterationID, ReqMap contains the continuation that takes the response to the request.

Constructors

ReqMap

Fields

unReqMap :: Map (RequestID, IterationID) (i -> Eff (ResumableEffs i o effs a) a)