Plutus.Contract.Secrets

Documentation

data Secret a

A secret value. A value of type `Secret a` can't leak onto the blockchain in plain-text unless you use an unsafe function. However, a value of type `Secret a` can end up on the blockchain via one of the escape hatches like escape_sha2_256.

data SecretArgument a

Secret argments are provided in the endpoint argument types.

This type guarantees that a `SecretArgument a` that is seen by the endpoint code is a `Secret a` in a way that can not be bypassed by safe code.

Constructors

UserSide a

EndpointSide (Secret a)

mkSecret :: a -> Secret a

Turn a public value into a secret value

secretArg :: a -> SecretArgument a

Construct a secret argument

extractSecret :: SecretArgument a -> Secret a

Extract a secret value from a secret argument

escape_sha2_256 :: Secret BuiltinByteString -> BuiltinByteString

Take the sha2_256 hash of a secret value. The result of this function can be used on the blockchain.

unsafe_escape_secret :: Secret a -> a

Warning: [Requires Review] An escape hatch is being created. This should only be used in trusted code.