Plutus.Contract.State

Contract state

Types for initialising and running instances of Contracts. The types and functions in this module are convenient wrappers around types and functions from Types, exposing an interface that is suitable for consumption by the PAB. In particular this means that insertAndUpdateContract has a single argument, and its argument & return types can be serialised to JSON easily.

To actually run a contract, follow this workflow:

Call initialiseContract to get the initial ContractResponse.

Look at the hooks of this value and generate an answer to one of them. This answer is a Response s value.

Call insertAndUpdateContract with a ContractRequest whose oldState field has the value of newState of the previous response, and whose event is the next answer (step 2).

Take the new ContractResponse and go back to step 2, until you get a response with no requests, or an error.

data Contract w (s :: Row *) e a

Contract w s e a is a contract with schema s, producing a value of type a or an error e. See note [Contract Schema].

data State w e

The state of a Contract, containing all responses that have been fed to it, and checkpoints that it produced.

Constructors

State

Fields

record :: Responses e

checkpoints :: CheckpointStore

observableState :: w

data ContractRequest w s

A request sent to a contract instance. It contains the previous State of the instance, and a Response to one of the requests of the instance.

Constructors

ContractRequest

Fields

oldState :: State w (CheckpointKey, s)

event :: Response s

data ContractResponse w e s h

A response produced by a contract instance. It contains the new State, the list of endpoints that can be called, logs produced by the contract, possibly an error message, and the accumulated observable state.

Constructors

ContractResponse

Fields

newState :: State w (CheckpointKey, s)

Serialised state of the contract (internal)

hooks :: [Request h]

Open requests that can be handled

logs :: [LogMessage Value]

Logs produced by the contract

lastLogs :: [LogMessage Value]

Logs produced in the last step

err :: Maybe e

Error that happened during contract execution

lastState :: w

Observable state produced in the last step

mapE :: forall e f w s h. (e -> f) -> ContractResponse w e s h -> ContractResponse w f s h

mapW :: forall w q e s h. (w -> q) -> ContractResponse w e s h -> ContractResponse q e s h

insertAndUpdateContract

Arguments

:: forall w s e a. Monoid w

=> Contract w s e a
The Contract with schema s error type e.

-> ContractRequest w PABResp
The ContractRequest value with the previous state and the new event.

-> ContractResponse w e PABResp PABReq

Run one step of the contract by restoring it to its previous state and feeding it a single new Response event.

initialiseContract :: forall w s e a. Monoid w => Contract w s e a -> ContractResponse w e PABResp PABReq

The ContractResponse with the initial state of the contract.

mkResponse :: forall w e s h a. Monoid w => w -> ResumableResult w e s h a -> ContractResponse w e s h