Plutus.Contract.StateMachine.ThreadToken

Description

Thread token data type definition and minting policy. Thread tokens are used to identify the contract instance on the blockchain, and ensuring that the state was produced by running the state machine from its initial state.

Documentation

data ThreadToken

Constructors

ThreadToken

Fields

ttOutRef :: TxOutRef

ttCurrencySymbol :: CurrencySymbol

checkPolicy :: TxOutRef -> (ValidatorHash, MintingPolarity) -> ScriptContext -> Bool

curPolicy :: TxOutRef -> MintingPolicy

threadTokenValue :: CurrencySymbol -> ValidatorHash -> Value

The Value containing exactly the thread token.

checkThreadTokenInner

Arguments

:: CurrencySymbol
The currency symbol of the thread token.

-> ValidatorHash
The hash of the (state machine) validator script using this thread token. This is used as the TokenName of the thread token.

-> Value
The value to check.

-> Integer
The expected number of thread tokens in the given value, n.

-> Bool
True if and only if exactly n thread tokens (and no other tokens) with the given CurrencySymbol are in the given Value.

Check exactly n thread tokens and no other tokens with the given CurrencySymbol are in the given Value.

checkThreadToken :: Maybe ThreadToken -> ValidatorHash -> Value -> Integer -> Bool