Plutus.Contract.Test

Description

Testing contracts with HUnit and Tasty

Documentation

module Plutus.Contract.Trace

newtype TracePredicateF a

Constructors

TracePredicate (forall effs. Members TestEffects effs => FoldM (Eff effs) EmulatorEvent a)

type TracePredicate = TracePredicateF Bool

type ContractConstraints s = (Forall (Output s) Unconstrained1, Forall (Input s) Unconstrained1, AllUniqueLabels (Input s), AllUniqueLabels (Output s), Forall (Input s) FromJSON, Forall (Input s) ToJSON, Forall (Output s) FromJSON, Forall (Output s) ToJSON)

not :: TracePredicate -> TracePredicate

(.&&.) :: TracePredicate -> TracePredicate -> TracePredicate infixl 3

(.||.) :: TracePredicate -> TracePredicate -> TracePredicate infixl 2

w1 :: Wallet

w2 :: Wallet

w3 :: Wallet

w4 :: Wallet

w5 :: Wallet

w6 :: Wallet

w7 :: Wallet

w8 :: Wallet

w9 :: Wallet

w10 :: Wallet

Assertions

endpointAvailable :: forall (l :: Symbol) w s e a. (KnownSymbol l, Monoid w) => Contract w s e a -> ContractInstanceTag -> TracePredicate

assertDone :: forall contract w s e a. (Monoid w, IsContract contract) => contract w s e a -> ContractInstanceTag -> (a -> Bool) -> String -> TracePredicate

A TracePredicate checking that the wallet's contract instance finished without errors.

assertNotDone :: forall contract w s e a. (Monoid w, IsContract contract) => contract w s e a -> ContractInstanceTag -> String -> TracePredicate

A TracePredicate checking that the wallet's contract instance is waiting for input.

assertContractError :: forall contract w s e a. (Monoid w, IsContract contract) => contract w s e a -> ContractInstanceTag -> (e -> Bool) -> String -> TracePredicate

A TracePredicate checking that the wallet's contract instance failed with an error.

data Outcome e a

Constructors

Done a
The contract finished without errors and produced a result

NotDone
The contract is waiting for more input.

Failed e
The contract failed with an error.

assertOutcome :: forall contract w s e a. (Monoid w, IsContract contract) => contract w s e a -> ContractInstanceTag -> (Outcome e a -> Bool) -> String -> TracePredicate

assertInstanceLog :: ContractInstanceTag -> ([EmulatorTimeEvent ContractInstanceLog] -> Bool) -> TracePredicate

assertNoFailedTransactions :: TracePredicate

Assert that no transaction failed to validate.

assertValidatedTransactionCount :: Int -> TracePredicate

Assert that n transactions validated, and no transaction failed to validate.

assertValidatedTransactionCountOfTotal :: Int -> Int -> TracePredicate

Assert that n transactions validated, and the rest failed.

assertFailedTransaction :: (CardanoTx -> ValidationError -> Bool) -> TracePredicate

Assert that at least one transaction failed to validate, and that all transactions that failed meet the predicate.

assertEvaluationError :: Text -> TracePredicate

Assert that at least one transaction failed to validate with an EvaluationError containing the given text.

assertHooks :: forall w s e a. Monoid w => Contract w s e a -> ContractInstanceTag -> ([PABReq] -> Bool) -> String -> TracePredicate

assertResponses :: forall w s e a. Monoid w => Contract w s e a -> ContractInstanceTag -> ([Response PABResp] -> Bool) -> String -> TracePredicate

Make an assertion about the responses provided to the contract instance.

assertUserLog :: ([EmulatorTimeEvent UserThreadMsg] -> Bool) -> TracePredicate

assertBlockchain :: ([Block] -> Bool) -> TracePredicate

An assertion about the blockchain

assertChainEvents :: ([ChainEvent] -> Bool) -> TracePredicate

An assertion about the chain events

assertChainEvents' :: ([ChainEvent] -> String) -> ([ChainEvent] -> Bool) -> TracePredicate

An assertion about the chain events with a custom error message

assertAccumState :: forall contract w s e a. (Monoid w, Show w, IsContract contract) => contract w s e a -> ContractInstanceTag -> (w -> Bool) -> String -> TracePredicate

Make an assertion about the accumulated state w of a contract instance.

data Shrinking

Constructors

DoShrink

DontShrink

assertResumableResult :: forall w s e a. (Monoid w, Show e, Show a, Show w) => Contract w s e a -> ContractInstanceTag -> Shrinking -> (ResumableResult w e PABResp PABReq a -> Bool) -> String -> TracePredicate

make an assertion about the ContractInstanceState of a contract instance

assertUnbalancedTx :: forall w s e a. Monoid w => Contract w s e a -> ContractInstanceTag -> (UnbalancedTx -> Bool) -> String -> TracePredicate

anyUnbalancedTx :: forall w s e a. Monoid w => Contract w s e a -> ContractInstanceTag -> TracePredicate

assertEvents :: forall w s e a. Monoid w => Contract w s e a -> ContractInstanceTag -> ([PABResp] -> Bool) -> String -> TracePredicate

walletFundsChangePlutus :: Wallet -> Value -> TracePredicate

Check that the funds in the wallet have changed by the given amount, exluding fees, using the Plutus Value type.

walletFundsChange :: Wallet -> Value -> TracePredicate

Check that the funds in the wallet have changed by the given amount, exluding fees.

walletFundsExactChange :: Wallet -> Value -> TracePredicate

Check that the funds in the wallet have changed by the given amount, including fees.

walletFundsAssetClassChange :: Wallet -> AssetId -> Integer -> TracePredicate

walletPaidFees :: Wallet -> Lovelace -> TracePredicate

waitingForSlot :: forall w s e a. Monoid w => Contract w s e a -> ContractInstanceTag -> Slot -> TracePredicate

valueAtAddress :: CardanoAddress -> (Value -> Bool) -> TracePredicate

Check that the funds at an address meet some condition.

plutusValueAtAddress :: CardanoAddress -> (Value -> Bool) -> TracePredicate

dataAtAddress :: forall d. FromData d => CardanoAddress -> ([d] -> Bool) -> TracePredicate

reasonable :: Validator -> Integer -> Assertion

Assert that the size of a Validator is below the maximum.

reasonable' :: (String -> IO ()) -> Validator -> Integer -> Assertion

Checking predicates

checkPredicate

Arguments

:: String
Descriptive name of the test

-> TracePredicate
The predicate to check

-> EmulatorTrace ()

-> TestTree

Check if the emulator trace meets the condition

checkPredicateCoverage

Arguments

:: String
Descriptive name of the test

-> CoverageRef

-> TracePredicate
The predicate to check

-> EmulatorTrace ()

-> TestTree

checkPredicateCoverageOptions

Arguments

:: CheckOptions
Options to use

-> String
Descriptive name of the test

-> CoverageRef

-> TracePredicate
The predicate to check

-> EmulatorTrace ()

-> TestTree

checkPredicateOptions

Arguments

:: CheckOptions
Options to use

-> String
Descriptive name of the test

-> TracePredicate
The predicate to check

-> EmulatorTrace ()

-> TestTree

A version of checkPredicate with configurable CheckOptions

checkPredicateGen :: GeneratorModel -> TracePredicate -> EmulatorTrace () -> Property

Check if the emulator trace meets the condition, using the GeneratorModel to generate initial transactions for the blockchain

checkPredicateGenOptions :: CheckOptions -> GeneratorModel -> TracePredicate -> EmulatorTrace () -> Property

A version of checkPredicateGen with configurable CheckOptions.

Note that the InitialChainState in the EmulatorConfig of the CheckOptions will be replaced with the mockchainInitialTxPool generated by the model.

checkPredicateInner

Arguments

:: forall m a. Monad m

=> CheckOptions

-> TracePredicate

-> EmulatorTrace a

-> (String -> m ())
Print out debug information in case of test failures TODO: can we generalize Bool here? We need it to get extractPropertyResult to work

-> (Bool -> m ())
assert

-> (CoverageData -> m ())

-> m (Either EmulatorErr a)

Evaluate a trace predicate on an emulator trace, printing out debug information and making assertions as we go.

checkPredicateInnerStream

Arguments

:: forall m a. Monad m

=> CheckOptions

-> TracePredicate

-> (forall effs. Stream (Of (LogMessage EmulatorEvent)) (Eff effs) (Either EmulatorErr a))

-> (String -> m ())
Print out debug information in case of test failures

-> (Bool -> m ())
assert

-> (CoverageData -> m ())

-> m (Either EmulatorErr a)

checkEmulatorFails

Arguments

:: String
Descriptive name of the test

-> CheckOptions

-> TracePredicate

-> EmulatorTrace ()
The trace that should fail

-> TestTree

Check if the emulator trace fails with the condition

data CheckOptions

Options for running the

Constructors

CheckOptions

Fields

_minLogLevel :: LogLevel

Minimum log level for emulator log messages to be included in the test output (printed if the test fails)

_emulatorConfig :: EmulatorConfig

defaultCheckOptions :: CheckOptions

minLogLevel :: Lens' CheckOptions LogLevel

emulatorConfig :: Lens' CheckOptions EmulatorConfig

changeInitialWalletValue :: Wallet -> (Value -> Value) -> CheckOptions -> CheckOptions

Modify the value assigned to the given wallet in the initial distribution.

increaseTransactionLimits :: CheckOptions -> CheckOptions

Set higher limits on transaction size and execution units. This can be used to work around MaxTxSizeUTxO and ExUnitsTooBigUTxO errors. Note that if you need this your Plutus script will probably not validate on Mainnet.

Etc

goldenPir :: FilePath -> CompiledCode a -> TestTree

Compare a golden PIR file to the provided CompiledCode.