Plutus.Contract.Test.ContractModel.CrashTolerance

Extending contract models with a model of

data WithCrashTolerance state

This derived state is used to derive a new ContractModel on top of the state contract model that also specifies how the contract(s) behave when contract instances crash and restart.

class ContractModel state => CrashTolerance state where

Minimal complete definition

restartArguments, available

Methods

crash :: SomeContractInstanceKey state -> Spec state ()

Specifiy what happens when a contract instance crashes

restart :: SomeContractInstanceKey state -> Spec state ()

Specify what happens when a contract instance is restarted

restartArguments :: ModelState state -> ContractInstanceKey state w s e p -> p

Specify the arguments to give to a restarted contract

available :: Action state -> [SomeContractInstanceKey state] -> Bool

Check if an action is available given a list of alive contract instances.