Plutus.Contract.Test.ContractModel.Internal

Description

This module provides a framework for testing Plutus contracts built on Test.QuickCheck. The testing is model based, so to test a contract you define a type modelling the state of the contract (or set of contracts) and provide an instance of the ContractModel class. This instance specifies what operations (Actions) the contract supports, how they interact with the model state, and how to execute them in the blockchain emulator (Plutus.Trace.Emulator). Tests are evaluated by running sequences of actions (random or user-specified) in the emulator and comparing the state of the blockchain to the model state at the end.

Test cases are written in the DL monad, which supports mixing fixed sequences of actions with random actions, making it easy to write properties like it is always possible to get all funds out of the contract.

Documentation

module Plutus.Contract.Test.ContractModel.Internal.ContractInstance

pattern UnderlyingAction :: Action s -> Action (WithInstances s)

pattern Unilateral :: Wallet -> Action (WithInstances s)

data CoverageOptions

Options for controlling coverage checking requirements

checkCoverage tells you whether or not to run the coverage checks at all.

`endpointCoverageEq instance endpointName` tells us what percentage of tests are required to include a call to the endpoint endpointName in the contract at `instance`.

coverIndex is the coverage index obtained from the CompiledCodeIn of the validator.

Constructors

CoverageOptions

Fields

_checkCoverage :: Bool

_endpointCoverageReq :: ContractInstanceTag -> String -> Double

_coverageIndex :: CoverageIndex

_coverageIORef :: Maybe (IORef CoverageData)

newtype WithInstances s

Constructors

WithInstances

Fields

withoutInstances :: s

type CheckableContractModel state = (RunModel state (SpecificationEmulatorTrace state), ContractInstanceModel state)

type EmulatorTraceWithInstances s = Eff (State (Handles s) ': EmulatorEffects)

type SpecificationEmulatorTrace s = Eff (Reader (Handles s) ': BaseEmulatorEffects)

contractHandle :: (ContractInstanceModel state, Typeable w, Typeable schema, Typeable err, Typeable params) => ContractInstanceKey state w schema err params -> RunMonad (SpecificationEmulatorTrace state) (ContractHandle w schema err)

activateWallets :: ContractInstanceModel state => (SymToken -> AssetId) -> [StartContract state] -> EmulatorTraceWithInstances state ()

instancesForOtherWallets :: Wallet -> Handles state -> [ContractInstanceId]

Used to freeze other wallets when checking a NoLockedFundsProof.

liftSpecificationTrace :: SpecificationEmulatorTrace s a -> EmulatorTraceWithInstances s a

makeTheHash :: HashAlgorithm crypto => BuiltinByteString -> Hash crypto stuff

delay :: Integer -> RunMonad (SpecificationEmulatorTrace state) ()

`delay n` delays emulator execution by n slots

data Whitelist

A whitelist entry tells you what final log entry prefixes are acceptable for a given error

Constructors

Whitelist

Fields

errorPrefixes :: Set Text

data NoLockedFundsProofLight model

Constructors

NoLockedFundsProofLight

Fields

nlfplMainStrategy :: DL (WithInstances model) ()

data NoLockedFundsProof model

A "proof" that you can always recover the funds locked by a contract. The first component is a strategy that from any state of the contract can get all the funds out. The second component is a strategy for each wallet that from the same state, shows how that wallet can recover the same (or bigger) amount as using the first strategy, without relying on any actions being taken by the other wallets.

For instance, in a two player game where each player bets some amount of funds and the winner gets the pot, there needs to be a mechanism for the players to recover their bid if the other player simply walks away (perhaps after realising the game is lost). If not, it won't be possible to construct a NoLockedFundsProof that works in a state where both players need to move before any funds can be collected.

Constructors

NoLockedFundsProof

Fields

nlfpMainStrategy :: DL (WithInstances model) ()

Strategy to recover all funds from the contract in any reachable state.

nlfpWalletStrategy :: Wallet -> DL (WithInstances model) ()

A strategy for each wallet to recover as much (or more) funds as the main strategy would give them in a given state, without the assistance of any other wallet.

nlfpOverhead :: ModelState (WithInstances model) -> SymValue

An initial amount of overhead value that may be lost - e.g. setup fees for scripts that can't be recovered.

nlfpErrorMargin :: ModelState (WithInstances model) -> SymValue

The total amount of margin for error in the value collected by the WalletStrategy compared to the MainStrategy. This is useful if your contract contains rounding code that makes the order of operations have a small but predictable effect on the value collected by different wallets.

checkCoverage :: Lens' CoverageOptions Bool

coverageIORef :: Lens' CoverageOptions (Maybe (IORef CoverageData))

coverageIndex :: Lens' CoverageOptions CoverageIndex

endpointCoverageReq :: Lens' CoverageOptions (ContractInstanceTag -> String -> Double)

defaultCoverageOptions :: CoverageOptions

Default coverage checking options are: * not to check coverage * set the requriements for every endpoint to 20% and * not to cover any source locations in the validator scripts.

defaultCheckOptionsContractModel :: CheckOptions

Default check options that include a large amount of Ada in the initial distributions to avoid having to write ContractModels that keep track of balances.

quickCheckWithCoverage :: Testable prop => Args -> CoverageOptions -> (CoverageOptions -> prop) -> IO CoverageReport

Run QuickCheck on a property that tracks coverage and print its coverage report.

quickCheckWithCoverageAndResult :: Testable prop => Args -> CoverageOptions -> (CoverageOptions -> prop) -> IO (CoverageReport, Result)

propRunActions_

Arguments

:: CheckableContractModel state

=> Actions (WithInstances state)
The actions to run

-> Property

Run a Actions in the emulator and check that the model and the emulator agree on the final wallet balance changes. Equivalent to

propRunActions_ hs actions = propRunActions hs (const $ pure True) actions

propRunActions

Arguments

:: CheckableContractModel state

=> (ModelState (WithInstances state) -> TracePredicate)
Predicate to check at the end

-> Actions (WithInstances state)
The actions to run

-> Property

Run a Actions in the emulator and check that the model and the emulator agree on the final wallet balance changes, and that the given TracePredicate holds at the end. Equivalent to:

propRunActions = propRunActionsWithOptions defaultCheckOptionsContractModel defaultCoverageOptions

propRunActionsWithOptions

Arguments

:: forall state. CheckableContractModel state

=> CheckOptions
Emulator options

-> CoverageOptions
Coverage options

-> (ModelState (WithInstances state) -> TracePredicate)
Predicate to check at the end of execution

-> Actions (WithInstances state)
The actions to run

-> Property

defaultNLFP :: NoLockedFundsProof model

The default skeleton of a NoLockedFundsProof - doesn't permit any overhead or error margin.

checkNoLockedFundsProof :: CheckableContractModel model => NoLockedFundsProof model -> Property

Check a NoLockedFundsProof. Each test will generate an arbitrary sequence of actions (anyActions_) and ask the nlfpMainStrategy to recover all funds locked by the contract after performing those actions. This results in some distribution of the contract funds to the wallets, and the test then asks each nlfpWalletStrategy to show how to recover their allotment of funds without any assistance from the other wallets (assuming the main strategy did not execute). When executing wallet strategies, the off-chain instances for other wallets are killed and their private keys are deleted from the emulator state.

checkNoLockedFundsProofFast :: CheckableContractModel model => NoLockedFundsProof model -> Property

checkNoLockedFundsProofWithOptions :: CheckableContractModel model => CheckOptions -> NoLockedFundsProof model -> Property

checkNoLockedFundsProofFastWithOptions :: CheckableContractModel model => CheckOptions -> NoLockedFundsProof model -> Property

checkNoLockedFundsProof' :: CheckableContractModel model => (Actions (WithInstances model) -> Property) -> NoLockedFundsProof model -> Property

actionsFromList :: [Action s] -> Actions s

varNumOf :: Act s -> Int

walletAddress :: Wallet -> AddressInEra Era

addressToWallet :: AddressInEra Era -> Maybe Wallet

checkNoLockedFundsProofLight :: CheckableContractModel model => NoLockedFundsProofLight model -> Property

isAcceptedBy :: Maybe Text -> Whitelist -> Bool

Check that the last entry in a log is accepted by a whitelist entry

whitelistOk :: Whitelist -> Bool

Check that a whitelist does not accept any partial functions

mkWhitelist :: [Text] -> Whitelist

defaultWhitelist :: Whitelist

checkErrorWhitelist :: CheckableContractModel m => Whitelist -> Actions (WithInstances m) -> Property

Check that running a contract model does not result in validation failures that are not accepted by the whitelist.

checkErrorWhitelistWithOptions :: forall m. CheckableContractModel m => CheckOptions -> CoverageOptions -> Whitelist -> Actions (WithInstances m) -> Property

Check that running a contract model does not result in validation failures that are not accepted by the whitelist.

Orphan instances

Member Waiting effs => IsRunnable (Eff effs)

Instance details

Methods

awaitSlot :: SlotNo -> Eff effs ()

HasChainIndex (EmulatorTraceWithInstances state)

Instance details

Methods

getChainIndex :: EmulatorTraceWithInstances state ChainIndex