Plutus.Contract.Test.ContractModel.Symbolics

Documentation

newtype AssetKey

Constructors

AssetKey Int

data SymToken

A symbolic token is a token that exists only during ContractModel generation time

Constructors

SymToken

Fields

symVar :: Var AssetKey

symVarIdx :: String

data SymValue

A symbolic value is a combination of a real value and a value associating symbolic tokens with an amount

Constructors

SymValue

Fields

symValMap :: Map SymToken Integer

actualValPart :: Value

symIsZero :: SymValue -> Bool

Check if a symbolic value is zero

symLeq :: SymValue -> SymValue -> Bool

Check if one symbolic value is less than or equal to another

toValue :: (SymToken -> AssetClass) -> SymValue -> Value

Using a semantics function for symbolic tokens, convert a SymValue to a Value

toSymVal :: (AssetClass -> Maybe SymToken) -> Value -> SymValue

Invert a sym token mapping to turn a Value into a SymValue, useful for error reporting

inv :: SymValue -> SymValue

class SymValueLike v where

Methods

toSymValue :: v -> SymValue

class TokenLike t where

Methods

symAssetIdValueOf :: SymValue -> t -> Integer

Get the value of a specific token in a SymValue

symAssetIdValue :: t -> Integer -> SymValue

Convert a token and an amount to a SymValue