Plutus.Contract.Test.Coverage

Documentation

getInvokedEndpoints :: [EmulatorEvent] -> Map ContractInstanceTag (Set String)

Get every endpoint name that has been invoked in the emulator events in es indexed by ContractInstanceTag

getCoverageData :: [EmulatorEvent] -> CoverageData

Collect every executed coverage annotation in the validators executed in es

newtype CoverageRef

Constructors

CoverageRef (IORef CoverageData)

newCoverageRef :: IO CoverageRef

readCoverageRef :: CoverageRef -> IO CoverageData

writeCoverageReport :: String -> CoverageReport -> IO ()

Write a coverage report to name.html for the given index.