Plutus.Trace.Emulator

Documentation

data Emulator

type EmulatorTrace = Eff EmulatorEffects

type EmulatorEffects = StartContract ': BaseEmulatorEffects

type BaseEmulatorEffects = [RunContract, Assert, Waiting, EmulatorControl, EmulatedWalletAPI, LogMsg String, Error EmulatorRuntimeError]

data EmulatorErr

Constructors

WalletErr WalletAPIError

ChainIndexErr ChainIndexError

AssertionErr AssertionError

InstanceErr EmulatorRuntimeError

ExitWasNeverCalled

data ContractHandle w s e

A reference to a running contract in the emulator.

Constructors

ContractHandle

Fields

chContract :: Contract w s e ()

chInstanceId :: ContractInstanceId

chInstanceTag :: ContractInstanceTag

chNetworkId :: NetworkId

data ContractInstanceTag

A user-defined tag for a contract instance. Used to find the instance's log messages in the emulator log.

type ContractConstraints s = (Forall (Output s) Unconstrained1, Forall (Input s) Unconstrained1, AllUniqueLabels (Input s), AllUniqueLabels (Output s), Forall (Input s) FromJSON, Forall (Input s) ToJSON, Forall (Output s) FromJSON, Forall (Output s) ToJSON)

Constructing Traces

assert :: forall effs. Member Assert effs => String -> (EmulatorState -> Bool) -> Eff effs ()

activateContract :: forall (contract :: Type -> Row Type -> Type -> Type -> Type) (s :: Row Type) e w a. (IsContract contract, ContractConstraints s, Show e, FromJSON e, ToJSON e, ToJSON w, Monoid w, FromJSON w) => forall effs. Member StartContract effs => Wallet -> contract w s e a -> ContractInstanceTag -> Eff effs (ContractHandle w s e)

activateContractWallet :: forall contract w s e effs. (IsContract contract, ContractConstraints s, Show e, ToJSON e, FromJSON e, ToJSON w, FromJSON w, Member StartContract effs, Monoid w) => Wallet -> contract w s e () -> Eff effs (ContractHandle w s e)

Like activateContract, but using walletInstanceTag for the tag.

walletInstanceTag :: Wallet -> ContractInstanceTag

The ContractInstanceTag for the contract instance of a wallet. See note [Wallet contract instances]

callEndpoint :: forall l ep w s e effs. (ToJSON ep, ContractConstraints s, HasEndpoint l ep s, Member RunContract effs) => ContractHandle w s e -> ep -> Eff effs ()

Call an endpoint on a contract instance.

getContractState :: forall w (s :: Row Type) e. (ContractConstraints s, FromJSON e, FromJSON w, ToJSON w) => forall effs. Member RunContract effs => ContractHandle w s e -> Eff effs (ContractInstanceState w s e ())

observableState :: forall w s e effs. (Member RunContract effs, ContractConstraints s, FromJSON e, FromJSON w, ToJSON w) => ContractHandle w s e -> Eff effs w

Get the observable state w of a contract instance.

activeEndpoints :: forall w s e effs. (Member RunContract effs, ContractConstraints s, FromJSON e, FromJSON w, ToJSON w) => ContractHandle w s e -> Eff effs [ActiveEndpoint]

Get the active endpoints of a contract instance.

liftWallet :: forall a. forall effs. Member EmulatedWalletAPI effs => Wallet -> Eff ('(:) WalletEffect ('(:) (Error WalletAPIError) ('(:) (LogMsg Text) ('(:) (LogMsg RequestHandlerLogMsg) ('[] :: [Type -> Type]))))) a -> Eff effs a

payToWallet :: forall effs. Member EmulatedWalletAPI effs => Wallet -> Wallet -> Value -> Eff effs TxId

Make a payment from one wallet to another

nextSlot :: Member Waiting effs => Eff effs Slot

Wait until the beginning of the next slot, returning the new slot number.

waitUntilSlot :: forall effs. Member Waiting effs => Slot -> Eff effs Slot

waitUntilTime :: Member Waiting effs => POSIXTime -> Eff effs POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

waitNSlots :: forall effs. Member Waiting effs => Natural -> Eff effs Slot

Wait for a number of slots

waitNMilliSeconds :: forall effs. Member Waiting effs => DiffMilliSeconds -> Eff effs Slot

Convert the given n milliseconds to a number of slots to wait.

Note: Currently, if n < length of a slot, then waitNMilliSeconds has no effect.

freezeContractInstance :: forall effs. Member EmulatorControl effs => ContractInstanceId -> Eff effs ()

thawContractInstance :: forall effs. Member EmulatorControl effs => ContractInstanceId -> Eff effs ()

Inspecting the chain state

setSigningProcess :: forall effs. Member EmulatorControl effs => Wallet -> Maybe SigningProcess -> Eff effs ()

chainState :: forall effs. Member EmulatorControl effs => Eff effs ChainState

getSlotConfig :: forall effs. Member EmulatorControl effs => Eff effs SlotConfig

chainNewestFirst :: Lens' ChainState Blockchain

txPool :: Lens' ChainState TxPool

index :: Lens' ChainState UtxoIndex

chainCurrentSlot :: Lens' ChainState Slot

Inspecting the agent states

agentState :: forall effs. Member EmulatorControl effs => Wallet -> Eff effs WalletState

ownPaymentPrivateKey :: WalletState -> PaymentPrivateKey

nodeClient :: Lens' WalletState NodeClientState

signingProcess :: Lens' WalletState (Maybe SigningProcess)

Throwing errors

throwError :: forall e (effs :: [Type -> Type]) a. Member (Error e) effs => e -> Eff effs a

data EmulatorRuntimeError

Constructors

ThreadIdNotFound ContractInstanceId

InstanceIdNotFound Wallet

EmulatorJSONDecodingError String Value

GenericError String

EmulatedWalletError WalletAPIError

AssertionError String

Running traces

data EmulatorConfig

Constructors

EmulatorConfig

Fields

_initialChainState :: InitialChainState

State of the blockchain at the beginning of the simulation. Can be given as a map of funds to wallets, or as a block of transactions.

_params :: Params

Set the protocol parameters, network ID and slot configuration for the emulator.

initialChainState :: Lens' EmulatorConfig InitialChainState

params :: Lens' EmulatorConfig Params

runEmulatorStream :: forall effs a. EmulatorConfig -> EmulatorTrace a -> Stream (Of (LogMessage EmulatorEvent)) (Eff effs) (Either EmulatorErr a, EmulatorState)

Run a 'Trace Emulator', streaming the log messages as they arrive

data TraceConfig

Options for how to set up and print the trace.

Constructors

TraceConfig

Fields

traceConfigShowEvent :: LogMessage EmulatorEvent -> Maybe String

Function to decide how to print the particular events.

traceConfigOutputHandle :: Handle

Where to print the outputs to. Default: stdout

traceConfigMinLogLevel :: LogLevel

traceConfigShowEventExample :: LogMessage EmulatorEvent -> Maybe String

Some example of how to configure the traceConfigShowEvent.

runEmulatorTrace :: TraceConfig -> EmulatorConfig -> EmulatorTrace a -> ([LogMessage EmulatorEvent], Either EmulatorErr a, EmulatorState)

Run an emulator trace to completion, returning a tuple of the final state of the emulator, the events, and any error, if any.

evalEmulatorTrace :: TraceConfig -> EmulatorConfig -> EmulatorTrace a -> Either EmulatorErr a

data PrintEffect r where

A very simple effect for interpreting the output printing done by the trace printing functions:

runEmulatorTraceEff

runEmulatorTraceIO

runEmulatorTraceIOWithConfig

Constructors

PrintLn :: String -> PrintEffect ()

runEmulatorTraceEff :: forall effs. Member PrintEffect effs => TraceConfig -> EmulatorConfig -> EmulatorTrace () -> Eff effs ()

Run the emulator trace returning an effect that can be evaluated by interpreting the PrintEffects.

runEmulatorTraceIO :: EmulatorTrace () -> IO ()

Runs the trace with runEmulatorTrace, with default configuration that prints a selection of events to stdout.

Example:

>>> runEmulatorTraceIO (void $ Trace.waitNSlots 1)

runEmulatorTraceIO' :: TraceConfig -> EmulatorConfig -> EmulatorTrace () -> IO ()

Deprecated: Renamed to runEmulatorTraceIOWithConfig

Interpreter

interpretEmulatorTrace :: forall effs a. (Member MultiAgentEffect effs, Member MultiAgentControlEffect effs, Member (Error EmulatorRuntimeError) effs, Member ChainControlEffect effs, Member (LogMsg EmulatorEvent') effs, Member (State EmulatorState) effs) => EmulatorConfig -> EmulatorTrace a -> Eff effs (Maybe a)

Interpret a 'Trace Emulator' action in the multi agent and emulated blockchain effects.