Plutus.Trace.Emulator.Extract

Documentation

data ValidatorMode

Constructors

FullyAppliedValidators

UnappliedValidators

writeScriptsTo

Arguments

:: ScriptsConfig
Configuration

-> String
Prefix to be used for file names

-> EmulatorTrace a
Emulator trace to extract transactions from

-> EmulatorConfig
Emulator config

-> IO (Sum Int64, ExBudget)

Run an emulator trace and write the applied scripts to a file in Flat format using the name as a prefix.

showStats :: Int64 -> ExBudget -> String

data ScriptsConfig

Configuration for writeScriptsTo

Constructors

ScriptsConfig

Fields

scPath :: FilePath

Folder the extracted scripts should be written to

scCommand :: Command

Whether to write out complete transactions or just the validator scripts

data Command

Command for writeScriptsTo

Constructors

Scripts
Write out validator scripts only (flat encoding)

Fields

unappliedValidators :: ValidatorMode

Whether to write fully applied or unapplied validators

Transactions
Write out partial transactions

Fields

networkId :: NetworkId

Network ID to use when creating addresses

protocolParamsJSON :: FilePath

Location of a JSON file with protocol parameters

MkTxLogs
Write out the arguments and results of mkTx calls