Wallet.API

Description

Mock wallet implementation

Documentation

data WalletEffect r

submitTxn :: forall effs. Member WalletEffect effs => CardanoTx -> Eff effs ()

ownPaymentPubKeyHash :: (Member WalletEffect effs, Member (Error WalletAPIError) effs) => Eff effs PaymentPubKeyHash

Deprecated: Use ownFirstPaymentPubKeyHash, ownPaymentPubKeyHashes or ownAddresses instead

ownPaymentPubKeyHashes :: Member WalletEffect effs => Eff effs [PaymentPubKeyHash]

ownFirstPaymentPubKeyHash :: (Member WalletEffect effs, Member (Error WalletAPIError) effs) => Eff effs PaymentPubKeyHash

ownAddresses :: forall effs. Member WalletEffect effs => Eff effs (NonEmpty CardanoAddress)

balanceTx :: forall effs. Member WalletEffect effs => UnbalancedTx -> Eff effs (Either WalletAPIError CardanoTx)

yieldUnbalancedTx :: forall effs. Member WalletEffect effs => UnbalancedTx -> Eff effs ()

data NodeClientEffect r

publishTx :: forall effs. Member NodeClientEffect effs => CardanoTx -> Eff effs ()

getClientSlot :: forall effs. Member NodeClientEffect effs => Eff effs Slot

getClientParams :: forall effs. Member NodeClientEffect effs => Eff effs Params

newtype PubKey

Constructors

PubKey

Fields

getPubKey :: LedgerBytes

newtype PubKeyHash

Constructors

PubKeyHash

Fields

getPubKeyHash :: BuiltinByteString

signTxAndSubmit :: Member WalletEffect effs => CardanoTx -> Eff effs CardanoTx

Add the wallet's signature to the transaction and submit it. Returns the transaction with the wallet's signature.

signTxAndSubmit_ :: Member WalletEffect effs => CardanoTx -> Eff effs ()

A version of signTxAndSubmit that discards the result.

payToAddress :: (Member WalletEffect effs, Member (Error WalletAPIError) effs, Member (LogMsg Text) effs, Member (LogMsg RequestHandlerLogMsg) effs) => Params -> SlotRange -> Value -> Address -> Eff effs CardanoTx

Transfer some funds to an address, returning the transaction that was submitted.

Note: Due to a constraint in the Cardano ledger, each tx output must have a minimum amount of Ada. Therefore, the funds to transfer will be adjusted to satisfy that constraint. See adjustUnbalancedTx.

payToAddress_ :: (Member WalletEffect effs, Member (Error WalletAPIError) effs, Member (LogMsg Text) effs, Member (LogMsg RequestHandlerLogMsg) effs) => Params -> SlotRange -> Value -> Address -> Eff effs ()

Transfer some funds to an address.

payToPaymentPublicKeyHash :: (Member WalletEffect effs, Member (Error WalletAPIError) effs, Member (LogMsg Text) effs, Member (LogMsg RequestHandlerLogMsg) effs) => Params -> SlotRange -> Value -> PaymentPubKeyHash -> Eff effs CardanoTx

Transfer some funds to an address locked by a public key, returning the transaction that was submitted.

Note: Due to a constraint in the Cardano ledger, each tx output must have a minimum amount of Ada. Therefore, the funds to transfer will be adjusted to satisfy that constraint. See adjustUnbalancedTx.

payToPaymentPublicKeyHash_ :: (Member WalletEffect effs, Member (Error WalletAPIError) effs, Member (LogMsg Text) effs, Member (LogMsg RequestHandlerLogMsg) effs) => Params -> SlotRange -> Value -> PaymentPubKeyHash -> Eff effs ()

Transfer some funds to an address locked by a public key.

data Params

Constructors

Params

Fields

pSlotConfig :: !SlotConfig

emulatorPParams :: !PParams

pNetworkId :: !NetworkId

Slot ranges

data Interval a

Constructors

Interval

Fields

ivFrom :: LowerBound a

ivTo :: UpperBound a

data Slot

type SlotRange = Interval Slot

width :: SlotRange -> Maybe Integer

defaultSlotRange :: SlotRange

The default slot validity range for transactions.

interval :: a -> a -> Interval a

singleton :: a -> Interval a

isEmpty :: (Enum a, Ord a) => Interval a -> Bool

always :: Interval a

member :: Ord a => a -> Interval a -> Bool

before :: Ord a => a -> Interval a -> Bool

after :: Ord a => a -> Interval a -> Bool

contains :: Ord a => Interval a -> Interval a -> Bool

Error handling

data WalletAPIError

An error thrown by wallet interactions.

Constructors

InsufficientFunds Text
There were insufficient funds to perform the desired operation.

ChangeHasLessThanNAda Value Ada
The change when selecting coins contains less than the minimum amount of Ada.

NoPaymentPubKeyHashError
The wallet doesn't have any payment key hash, which should not be possible.

PaymentPrivateKeyNotFound PaymentPubKeyHash
The private key of this public key hash is not known to the wallet.

ValidationError ValidationError
There was an error during off-chain validation.

ToCardanoError ToCardanoError
There was an error while converting to Cardano.API format.

PaymentMkTxError MkTxError
There was an error while creating a payment transaction

RemoteClientFunctionNotYetSupported Text
The called wallet effect is not yet supported in a remote wallet client scenario.

OtherError Text
Some other error occurred.

throwInsufficientFundsError :: Member (Error WalletAPIError) effs => Text -> Eff effs a

throwOtherError :: Member (Error WalletAPIError) effs => Text -> Eff effs a