Wallet.Effects

Wallet effect

data WalletEffect r where

Constructors

SubmitTxn :: CardanoTx -> WalletEffect ()

OwnAddresses :: WalletEffect (NonEmpty CardanoAddress)

BalanceTx :: UnbalancedTx -> WalletEffect (Either WalletAPIError CardanoTx)

TotalFunds

Deprecated: We won't use the wallet for querying blockchain information. See https://plutus-apps.readthedocs.io/en/latest/adr/0005-pab-indexing-solution-integration.html

Fields

:: WalletEffect Value

Total of all funds that are in the wallet (incl. tokens)

WalletAddSignature :: CardanoTx -> WalletEffect CardanoTx

YieldUnbalancedTx :: UnbalancedTx -> WalletEffect ()
Sends an unbalanced tx to be balanced, signed and submitted.

submitTxn :: forall effs. Member WalletEffect effs => CardanoTx -> Eff effs ()

ownAddresses :: forall effs. Member WalletEffect effs => Eff effs (NonEmpty CardanoAddress)

balanceTx :: forall effs. Member WalletEffect effs => UnbalancedTx -> Eff effs (Either WalletAPIError CardanoTx)

totalFunds :: forall effs. Member WalletEffect effs => Eff effs Value

walletAddSignature :: forall effs. Member WalletEffect effs => CardanoTx -> Eff effs CardanoTx

yieldUnbalancedTx :: forall effs. Member WalletEffect effs => UnbalancedTx -> Eff effs ()

Node client

data NodeClientEffect r where

Constructors

PublishTx :: CardanoTx -> NodeClientEffect ()

GetClientSlot :: NodeClientEffect Slot

GetClientParams :: NodeClientEffect Params

publishTx :: forall effs. Member NodeClientEffect effs => CardanoTx -> Eff effs ()

getClientSlot :: forall effs. Member NodeClientEffect effs => Eff effs Slot

getClientParams :: forall effs. Member NodeClientEffect effs => Eff effs Params