Wallet.Emulator.Folds

Documentation

type EmulatorEventFold a = Fold EmulatorEvent a

type EmulatorEventFoldM effs a = FoldM (Eff effs) EmulatorEvent a

A fold over emulator events that can fail with EmulatorFoldErr

data EmulatorFoldErr

Constructors

InstanceStateJSONDecodingError String (Response Value)

describeError :: EmulatorFoldErr -> String

A human-readable explanation of the error, to be included in the logs.

Folds for contract instances

instanceState :: forall w s e a effs. (Member (Error EmulatorFoldErr) effs, Monoid w) => Contract w s e a -> ContractInstanceTag -> EmulatorEventFoldM effs (Maybe (ContractInstanceState w s e a))

The state of a contract instance, recovered from the emulator log.

instanceRequests :: forall w s e a effs. (Member (Error EmulatorFoldErr) effs, Monoid w) => Contract w s e a -> ContractInstanceTag -> EmulatorEventFoldM effs [Request PABReq]

The list of open requests of the contract instance at its latest iteration

instanceResponses :: forall w s e a effs. (Member (Error EmulatorFoldErr) effs, Monoid w) => Contract w s e a -> ContractInstanceTag -> EmulatorEventFoldM effs [Response PABResp]

The reponses received by the contract instance

instanceOutcome :: forall w s e a effs. (Member (Error EmulatorFoldErr) effs, Monoid w) => Contract w s e a -> ContractInstanceTag -> EmulatorEventFoldM effs (Outcome e a)

The final state of the instance

instanceTransactions :: forall w s e a effs. (Member (Error EmulatorFoldErr) effs, Monoid w) => Contract w s e a -> ContractInstanceTag -> EmulatorEventFoldM effs [UnbalancedTx]

The unbalanced transactions generated by the contract instance.

data Outcome e a

Constructors

Done a
The contract finished without errors and produced a result

NotDone
The contract is waiting for more input.

Failed e
The contract failed with an error.

instanceLog :: ContractInstanceTag -> EmulatorEventFold [EmulatorTimeEvent ContractInstanceLog]

The log messages produced by the contract instance.

instanceAccumState :: forall w s e a effs. (Member (Error EmulatorFoldErr) effs, Monoid w) => Contract w s e a -> ContractInstanceTag -> EmulatorEventFoldM effs w

Accumulated state of the contract instance

Folds for transactions and the UTXO set

chainEvents :: EmulatorEventFold [ChainEvent]

All chain events emitted by the node

failedTransactions :: Maybe ValidationPhase -> EmulatorEventFold [(CardanoTx, ValidationError, Value)]

Transactions that failed to validate, in the given validation phase (if specified).

validatedTransactions :: EmulatorEventFold [(OnChainTx, RedeemerReport)]

Transactions that were validated

utxoAtAddress :: CardanoAddress -> EmulatorEventFold UtxoMap

Unspent outputs at an address

valueAtAddress :: CardanoAddress -> EmulatorEventFold Value

The total value of unspent outputs at an address

Folds for individual wallets (emulated agents)

walletFunds :: Wallet -> EmulatorEventFold Value

The funds belonging to a wallet

walletFees :: Wallet -> EmulatorEventFold Lovelace

The fees paid by a wallet

walletTxBalanceEvents :: EmulatorEventFold [UnbalancedTx]

Unbalanced transactions that are sent to the wallet for balancing

walletsAdjustedTxEvents :: EmulatorEventFold [Lovelace]

Min lovelace of txOuts from adjusted unbalanced transactions for all wallets

Folds that are used in the Playground

annotatedBlockchain :: EmulatorEventFold [[AnnotatedTx]]

Annotate the transactions that were validated by the node

blockchain :: EmulatorEventFold [Block]

All transactions that happened during the simulation

emulatorLog :: EmulatorEventFold [EmulatorEvent]

The list of all emulator events

userLog :: EmulatorEventFold [EmulatorTimeEvent UserThreadMsg]

Log and error messages produced by the main (user) thread in the emulator

Etc.

renderLines :: forall a. Pretty a => Fold a Text

Pretty-print each element into a new line.

preMapMaybeM :: Monad m => (a -> m (Maybe b)) -> FoldM m b r -> FoldM m a r

An effectful mapMaybe for FoldM.

preMapMaybe :: (a -> Maybe b) -> Fold b r -> Fold a r

mapMaybe for Fold.

postMapM :: Monad m => (b -> m c) -> FoldM m a b -> FoldM m a c

Effectfully map the result of a FoldM

mkTxLogs :: EmulatorEventFold [MkTxLog]