Wallet.Types

Description

Defines a number of types that are used in Wallet.XXX modules

Documentation

newtype ContractInstanceId

Unique ID for contract instance

Constructors

ContractInstanceId

Fields

unContractInstanceId :: UUID

contractInstanceIDs :: [ContractInstanceId]

A pure list of all ContractInstanceId values. To be used in testing.

randomID :: IO ContractInstanceId

data ContractActivityStatus

Constructors

Active

Stopped

Done

parseContractActivityStatus :: Text -> Maybe ContractActivityStatus

data Notification

Constructors

Notification

Fields

notificationContractID :: ContractInstanceId

notificationContractEndpoint :: EndpointDescription

notificationContractArg :: Value

data NotificationError

Constructors

EndpointNotAvailable ContractInstanceId EndpointDescription

MoreThanOneEndpointAvailable ContractInstanceId EndpointDescription

InstanceDoesNotExist ContractInstanceId

NotificationJSONDecodeError EndpointDescription Value String
Indicates that the target contract does not have the expected schema

TODO: SCP-2137 Not currently used. As endpoint parameter decoding happends inside the Contract and a throwError is used is decoding failed. However, still valuable to be used by the PAB to throw an error is an endpoint could not be decoded.

newtype EndpointDescription

Constructors

EndpointDescription

Fields

getEndpointDescription :: String

newtype EndpointValue a

Constructors

EndpointValue

Fields

unEndpointValue :: a