Plutus.Contract.Error

Documentation

data ContractError

Constructors

WalletContractError WalletAPIError

ChainIndexContractError Text ChainIndexResponse

EmulatorAssertionContractError AssertionError

ConstraintResolutionContractError MkTxError

TxConstraintResolutionContractError MkTxError

TxToCardanoConvertContractError ToCardanoError

ResumableContractError MatchingError

CCheckpointContractError CheckpointError

EndpointDecodeContractError

Fields

eeEndpointDescription :: EndpointDescription

The endpoint description which the decoding error occurred from

eeEndpointValue :: EndpointValue Value

The endpoint value that was used as an endpoint parameter

eeErrorMessage :: Text

JSON decoding error message

OtherContractError Text

class AsContractError r where

Minimal complete definition

_ContractError

Methods

_ContractError :: Prism' r ContractError

_WalletContractError :: Prism' r WalletAPIError

_ChainIndexContractError :: Prism' r (Text, ChainIndexResponse)

_EmulatorAssertionContractError :: Prism' r AssertionError

_ConstraintResolutionContractError :: Prism' r MkTxError

_TxConstraintResolutionContractError :: Prism' r MkTxError

_TxToCardanoConvertContractError :: Prism' r ToCardanoError

_ResumableContractError :: Prism' r MatchingError

_CCheckpointContractError :: Prism' r CheckpointError

_EndpointDecodeContractError :: Prism' r (EndpointDescription, EndpointValue Value, Text)

_OtherContractError :: Prism' r Text

newtype MatchingError

An error

Constructors

WrongVariantError

Fields

unWrongVariantError :: Text

class AsMatchingError r where

Minimal complete definition

_MatchingError

Methods

_MatchingError :: Prism' r MatchingError

_WrongVariantError :: Prism' r Text

newtype AssertionError

An error emitted when an Assertion fails.

Constructors

GenericAssertion

Fields

unAssertionError :: Text

class AsAssertionError r where

Minimal complete definition

_AssertionError

Methods

_AssertionError :: Prism' r AssertionError

_GenericAssertion :: Prism' r Text