Plutus.Contract.Request

PAB requests

Waiting

awaitSlot :: forall w s e. AsContractError e => Slot -> Contract w s e Slot

Wait until the slot

isSlot :: forall w s e. AsContractError e => Slot -> Promise w s e Slot

Wait until the slot

currentSlot :: forall w s e. AsContractError e => Contract w s e Slot

Deprecated: Use currentNodeClientSlot instead

Get the current slot number

currentPABSlot :: forall w s e. AsContractError e => Contract w s e Slot

Deprecated: Use currentNodeClientSlot instead

Get the current slot number of PAB

currentNodeClientSlot :: forall w s e. AsContractError e => Contract w s e Slot

Get the current slot number of the node client (the local or remote node) that the application is connected to.

currentChainIndexSlot :: forall w s e. AsContractError e => Contract w s e Slot

Get the current node slot number querying slot number from plutus chain index to be aligned with slot at local running node

waitNSlots :: forall w s e. AsContractError e => Natural -> Contract w s e Slot

Wait for a number of slots to pass

awaitTime :: forall w s e. AsContractError e => POSIXTime -> Contract w s e POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

Example: if starting time is 0 and slot length is 3s, then `awaitTime 4` waits until slot 2 and returns the value `POSIXTime 5`.

isTime :: forall w s e. AsContractError e => POSIXTime -> Promise w s e POSIXTime

Wait until the slot where the given time falls into and return latest time we know has passed.

currentTime :: forall w s e. AsContractError e => Contract w s e POSIXTime

Deprecated: Use currentNodeClientTimeRange instead

Get the latest time of the current slot.

Example: if slot length is 3s and current slot is 2, then currentTime returns the value `POSIXTime 5`

currentNodeClientTimeRange :: forall w s e. AsContractError e => Contract w s e (POSIXTime, POSIXTime)

Get the POSIXTime range of the current slot.

Example: if slot length is 3s and current slot is 2, then currentTimeRange returns the time interval [3, 5[.

waitNMilliSeconds :: forall w s e. AsContractError e => DiffMilliSeconds -> Contract w s e POSIXTime

Wait for a number of milliseconds starting at the ending time of the current slot, and return the latest time we know has passed.

Example: if starting time is 0, slot length is 3000ms and current slot is 0, then `waitNMilliSeconds 0` returns the value `POSIXTime 2000` and `waitNMilliSeconds 1000` returns the value `POSIXTime 5`.

Chain index queries

datumFromHash :: forall w s e. AsContractError e => DatumHash -> Contract w s e (Maybe Datum)

datumsAt :: forall w s e. AsContractError e => CardanoAddress -> Contract w s e [Datum]

Get the all datums at an address whether or not the corresponding utxo have been consumed or not.

validatorFromHash :: forall w s e. AsContractError e => ValidatorHash -> Contract w s e (Maybe (Versioned Validator))

mintingPolicyFromHash :: forall w s e. AsContractError e => MintingPolicyHash -> Contract w s e (Maybe (Versioned MintingPolicy))

stakeValidatorFromHash :: forall w s e. AsContractError e => StakeValidatorHash -> Contract w s e (Maybe (Versioned StakeValidator))

redeemerFromHash :: forall w s e. AsContractError e => RedeemerHash -> Contract w s e (Maybe Redeemer)

txOutFromRef :: forall w s e. AsContractError e => TxOutRef -> Contract w s e (Maybe DecoratedTxOut)

txFromTxId :: forall w s e. AsContractError e => TxId -> Contract w s e (Maybe ChainIndexTx)

unspentTxOutFromRef :: forall w s e. AsContractError e => TxOutRef -> Contract w s e (Maybe DecoratedTxOut)

utxoRefMembership :: forall w s e. AsContractError e => TxOutRef -> Contract w s e IsUtxoResponse

utxoRefsAt :: forall w s e. AsContractError e => PageQuery TxOutRef -> CardanoAddress -> Contract w s e UtxosResponse

Get the unspent transaction output references at an address.

utxoRefsWithCurrency :: forall w s e. AsContractError e => PageQuery TxOutRef -> AssetClass -> Contract w s e UtxosResponse

Get the unspent transaction output references with a specific currrency (AssetClass).

utxosAt :: forall w s e. AsContractError e => CardanoAddress -> Contract w s e (Map TxOutRef DecoratedTxOut)

Get the unspent transaction outputs at an address.

utxosTxOutTxFromTx :: AsContractError e => ChainIndexTx -> Contract w s e [(TxOutRef, (DecoratedTxOut, ChainIndexTx))]

Get the unspent transaction outputs from a ChainIndexTx.

utxosTxOutTxAt :: forall w s e. AsContractError e => CardanoAddress -> Contract w s e (Map TxOutRef (DecoratedTxOut, ChainIndexTx))

Get unspent transaction outputs with transaction from address.

txsFromTxIds :: forall w s e. AsContractError e => [TxId] -> Contract w s e [ChainIndexTx]

Get the transactions for a list of transaction ids.

txoRefsAt :: forall w s e. AsContractError e => PageQuery TxOutRef -> CardanoAddress -> Contract w s e TxosResponse

Get the transaction outputs at an address.

txsAt :: forall w s e. AsContractError e => CardanoAddress -> Contract w s e [ChainIndexTx]

Get the transactions at an address.

getTip :: forall w s e. AsContractError e => Contract w s e Tip

collectQueryResponse

Arguments

:: Monad m

=> (PageQuery TxOutRef -> m (QueryResponse a))
query response function

-> m [a]

Go through each Pages of QueryResponse, and collect the results.

Waiting for changes to the UTXO set

fundsAtAddressGt :: forall w s e. AsContractError e => CardanoAddress -> Value -> Contract w s e (Map TxOutRef DecoratedTxOut)

Watch an address for changes, and return the outputs at that address when the total value at the address has surpassed the given value.

fundsAtAddressGeq :: forall w s e. AsContractError e => CardanoAddress -> Value -> Contract w s e (Map TxOutRef DecoratedTxOut)

Watch an address for changes, and return the outputs at that address when the total value at the address has reached or surpassed the given value.

fundsAtAddressCondition :: forall w s e. AsContractError e => (Value -> Bool) -> CardanoAddress -> Contract w s e (Map TxOutRef DecoratedTxOut)

watchAddressUntilSlot :: forall w s e. AsContractError e => CardanoAddress -> Slot -> Contract w s e (Map TxOutRef DecoratedTxOut)

Wait until the target slot and get the unspent transaction outputs at an address.

watchAddressUntilTime :: forall w s e. AsContractError e => CardanoAddress -> POSIXTime -> Contract w s e (Map TxOutRef DecoratedTxOut)

Wait until the target time and get the unspent transaction outputs at an address.

awaitUtxoSpent :: forall w s e. AsContractError e => TxOutRef -> Contract w s e ChainIndexTx

Wait until the UTXO has been spent, returning the transaction that spends it.

utxoIsSpent :: forall w s e. AsContractError e => TxOutRef -> Promise w s e ChainIndexTx

Wait until the UTXO has been spent, returning the transaction that spends it.

awaitUtxoProduced :: forall w s e. AsContractError e => CardanoAddress -> Contract w s e (NonEmpty ChainIndexTx)

Wait until one or more unspent outputs are produced at an address.

utxoIsProduced :: forall w s e. AsContractError e => CardanoAddress -> Promise w s e (NonEmpty ChainIndexTx)

Wait until one or more unspent outputs are produced at an address.

Tx and tx output confirmation

data RollbackState a

The rollback state of a Cardano transaction

Constructors

Unknown
The transaction is not on the chain. That's all we can say.

type TxStatus = RollbackState ()

The status of a Cardano transaction

awaitTxStatusChange :: forall w s e. AsContractError e => TxId -> Contract w s e TxStatus

Wait for the status of a transaction to change

awaitTxConfirmed :: forall w s e. AsContractError e => TxId -> Contract w s e ()

Wait until a transaction is confirmed (added to the ledger). If the transaction is never added to the ledger then awaitTxConfirmed never returns

isTxConfirmed :: forall w s e. AsContractError e => TxId -> Promise w s e ()

Wait until a transaction is confirmed (added to the ledger).

type TxOutStatus = RollbackState TxOutState

awaitTxOutStatusChange :: forall w s e. AsContractError e => TxOutRef -> Contract w s e TxOutStatus

Wait for the status of a transaction output to change.

Contract instances

ownInstanceId :: forall w s e. AsContractError e => Contract w s e ContractInstanceId

Get the ContractInstanceId of this instance.

Exposing endpoints

type HasEndpoint l a s = (HasType l (EndpointValue a) (Input s), HasType l ActiveEndpoint (Output s), KnownSymbol l, ContractRow s)

newtype EndpointDescription

Constructors

EndpointDescription String

type Endpoint l a = l .== (EndpointValue a, ActiveEndpoint)

endpoint :: forall l a w s e b. (HasEndpoint l a s, AsContractError e, FromJSON a) => (a -> Contract w s e b) -> Promise w s e b

Expose an endpoint, return the data that was entered

handleEndpoint :: forall l a w s e1 e2 b. (HasEndpoint l a s, AsContractError e1, FromJSON a) => (Either e1 a -> Contract w s e2 b) -> Promise w s e2 b

endpointWithMeta :: forall l a w s e meta b. (HasEndpoint l a s, AsContractError e, ToJSON meta, FromJSON a) => meta -> (a -> Contract w s e b) -> Promise w s e b

Expose an endpoint with some metadata. Return the data that was entered.

endpointDescription :: forall l. KnownSymbol l => Proxy l -> EndpointDescription

endpointReq :: forall l a s. HasEndpoint l a s => ActiveEndpoint

endpointResp :: forall l a s. (HasEndpoint l a s, ToJSON a) => a -> PABResp

Wallet information

ownPaymentPubKeyHash :: forall w s e. AsContractError e => Contract w s e PaymentPubKeyHash

Deprecated: Use ownFirstPaymentPubKeyHash, ownPaymentPubKeyHashes or ownAddresses instead

Get the hash of a public key belonging to the wallet that runs this contract. * Any funds paid to this public key hash will be treated as the wallet's own funds * The wallet is able to sign transactions with the private key of this public key, for example, if the public key is added to the requiredSignatures field of Tx. * There is a 1-n relationship between wallets and public keys (although in the mockchain n=1)

ownPaymentPubKeyHashes :: forall w s e. AsContractError e => Contract w s e [PaymentPubKeyHash]

ownFirstPaymentPubKeyHash :: forall w s e. AsContractError e => Contract w s e PaymentPubKeyHash

ownAddresses :: forall w s e. AsContractError e => Contract w s e (NonEmpty CardanoAddress)

Get the addresses belonging to the wallet that runs this contract. * Any funds paid to one of these addresses will be treated as the wallet's own funds * The wallet is able to sign transactions with the private key of one of its public key, for example, if the public key is added to the requiredSignatures field of Tx. * There is a 1-n relationship between wallets and addresses (although in the mockchain n=1)

ownAddress :: forall w s e. AsContractError e => Contract w s e CardanoAddress

Get the first address of the wallet that runs this contract.

ownUtxos :: forall w s e. AsContractError e => Contract w s e (Map TxOutRef DecoratedTxOut)

Get all utxos belonging to the wallet that runs this contract.

Submitting transactions

adjustUnbalancedTx :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e UnbalancedTx

Adjust the unbalanced tx

submitUnbalancedTx :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e CardanoTx

Send an unbalanced transaction to be balanced and signed. Returns the ID of the final transaction when the transaction was submitted. Throws an error if balancing or signing failed.

submitBalancedTx :: forall w s e. AsContractError e => CardanoTx -> Contract w s e CardanoTx

Send an balanced transaction to be signed. Returns the ID of the final transaction when the transaction was submitted. Throws an error if signing failed.

balanceTx :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e CardanoTx

Send an unbalanced transaction to be balanced. Returns the balanced transaction. Throws an error if balancing failed.

submitTx :: forall w s e. AsContractError e => TxConstraints Void Void -> Contract w s e CardanoTx

Build a transaction that satisfies the constraints, then submit it to the network. The constraints do not refer to any typed script inputs or outputs.

submitTxConstraints :: forall a w s e. (ToData (RedeemerType a), FromData (DatumType a), ToData (DatumType a), AsContractError e) => TypedValidator a -> TxConstraints (RedeemerType a) (DatumType a) -> Contract w s e CardanoTx

Build a transaction that satisfies the constraints, then submit it to the network. Using the current outputs at the contract address and the contract's own public key to solve the constraints.

submitTxConstraintsSpending :: forall a w s e. (ToData (RedeemerType a), FromData (DatumType a), ToData (DatumType a), AsContractError e) => TypedValidator a -> Map TxOutRef DecoratedTxOut -> TxConstraints (RedeemerType a) (DatumType a) -> Contract w s e CardanoTx

Build a transaction that satisfies the constraints using the UTXO map to resolve any input constraints (see InputConstraint)

submitTxConstraintsWith :: forall a w s e. (ToData (RedeemerType a), FromData (DatumType a), ToData (DatumType a), AsContractError e) => ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> Contract w s e CardanoTx

Build a transaction that satisfies the constraints, then submit it to the network. Using the given constraints.

submitTxConfirmed :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e ()

A version of submitTx that waits until the transaction has been confirmed on the ledger before returning.

mkTxConstraints :: forall a w s e. (ToData (RedeemerType a), FromData (DatumType a), ToData (DatumType a), AsContractError e) => ScriptLookups a -> TxConstraints (RedeemerType a) (DatumType a) -> Contract w s e UnbalancedTx

Build a transaction that satisfies the constraints

yieldUnbalancedTx :: forall w s e. AsContractError e => UnbalancedTx -> Contract w s e ()

Take an UnbalancedTx then balance, sign and submit it to the blockchain without returning any results.

Parameters

getParams :: forall w s e. AsContractError e => Contract w s e Params

Get the configured parameter set.

Etc.

type ContractRow s = (AllUniqueLabels (Input s), AllUniqueLabels (Output s))

Constraints on the contract schema, ensuring that the labels of the schema are unique.

pabReq

Arguments

:: forall w s e a. AsContractError e

=> PABReq
The request to send

-> Prism' PABResp a
Prism for the response

-> Contract w s e a

data MkTxLog

Arguments and result of a call to mkTx

Constructors

MkTxLog

Fields

mkTxLogLookups :: ScriptLookups Any

mkTxLogTxConstraints :: TxConstraints BuiltinData BuiltinData

mkTxLogResult :: Either MkTxError UnbalancedTx