Plutus.Contract.StateMachine

Documentation

To write your contract as a state machine you need * Two types state and input for the state and inputs of the machine * A 'SM.StateMachineInstance state input' describing the transitions and checks of the state machine (this is the on-chain code) * A 'StateMachineClient state input' with the state machine instance and an allocation function

In many cases it is enough to define the transition function t :: (state, Value) -> input -> Maybe (TxConstraints state) and use mkStateMachine and mkStateMachineClient to get the client. You can then use runInitialise and runStep to initialise and transition the state machine. runStep gets the current state from the utxo set and makes the transition to the next state using the given input and taking care of all payments.

data StateMachineClient s i

Client-side definition of a state machine.

Constructors

StateMachineClient

Fields

scInstance :: StateMachineInstance s i

The instance of the state machine, defining the machine's transitions, its final states and its check function.

scChooser :: [OnChainState s i] -> Either SMContractError (OnChainState s i)

A function that chooses the relevant on-chain state, given a list of all potential on-chain states found at the contract address.

data TxConstraints i o

Restrictions placed on the allocation of funds to outputs of transactions.

data SMContractError

Constructors

ChooserError Text

UnableToExtractTransition

SMCContractError ContractError

class AsSMContractError r where

Minimal complete definition

_SMContractError

Methods

_SMContractError :: Prism' r SMContractError

_ChooserError :: Prism' r Text

_UnableToExtractTransition :: Prism' r ()

_SMCContractError :: Prism' r ContractError

data StateMachine s i

Specification of a state machine, consisting of a transition function that determines the next state from the current state and an input, and a checking function that checks the validity of the transition in the context of the current transaction.

Constructors

StateMachine

Fields

smTransition :: State s -> i -> Maybe (TxConstraints Void Void, State s)

The transition function of the state machine. Nothing indicates an invalid transition from the current state.

smFinal :: s -> Bool

Check whether a state is the final state

smCheck :: s -> i -> ScriptContext -> Bool

The condition checking function. Can be used to perform checks on the pending transaction that aren't covered by the constraints. smCheck is always run in addition to checking the constraints, so the default implementation always returns true.

smThreadToken :: Maybe ThreadToken

The ThreadToken that identifies the contract instance. Make one with getThreadToken and pass it on to mkStateMachine. Initialising the machine will then mint a thread token value.

data StateMachineInstance s i

Constructors

StateMachineInstance

Fields

stateMachine :: StateMachine s i

The state machine specification.

typedValidator :: TypedValidator (StateMachine s i)

The validator code for this state machine.

data State s

Constructors

State

Fields

stateData :: s

stateValue :: Value

newtype OnChainState s i

Typed representation of the on-chain state of a state machine instance

Constructors

OnChainState

Fields

ocsTxOutRef :: TypedScriptTxOutRef (StateMachine s i)

Typed UTXO

data WaitingResult t i s

The outcome of waitForUpdateTimeout

Constructors

Timeout t
The timeout happened before any change of the on-chain state was detected

ContractEnded i
The state machine instance ended

Transition i s
The state machine instance transitioned to a new state

InitialState s
The state machine instance was initialised

data InvalidTransition s i

An invalid transition

Constructors

InvalidTransition

Fields

tfState :: Maybe (State s)

Current state. Nothing indicates that there is no current state.

tfInput :: i

Transition that was attempted but failed

data TransitionResult s i

Result of an attempted transition

Constructors

TransitionFailure (InvalidTransition s i)
The transition is not allowed

TransitionSuccess s
The transition is allowed and results in a new state

data ThreadToken

Constructors

ThreadToken TxOutRef CurrencySymbol

Constructing the machine instance

mkValidator :: forall s i. ToData s => StateMachine s i -> ValidatorType (StateMachine s i)

Turn a state machine into a validator script.

mkStateMachine :: Maybe ThreadToken -> (State s -> i -> Maybe (TxConstraints Void Void, State s)) -> (s -> Bool) -> StateMachine s i

A state machine that does not perform any additional checks on the ScriptContext (beyond enforcing the constraints)

Constructing the state machine client

mkStateMachineClient :: forall state input. StateMachineInstance state input -> StateMachineClient state input

A state machine client with the defaultChooser function

defaultChooser :: forall state input. [OnChainState state input] -> Either SMContractError (OnChainState state input)

A state chooser function that fails if confronted with anything other than exactly one output

getStates :: forall s i. (FromData s, ToData s) => StateMachineInstance s i -> Map TxOutRef DecoratedTxOut -> [OnChainState s i]

Running the state machine

runGuardedStep

Arguments

:: forall w a e state schema input. (AsSMContractError e, FromData state, ToData state, ToData input)

=> StateMachineClient state input
The state machine

-> input
The input to apply to the state machine

-> (UnbalancedTx -> state -> state -> Maybe a)
The guard to check before running the step

-> Contract w schema e (Either a (TransitionResult state input))

Tries to run one step of a state machine: If the guard (the last argument) returns Nothing when given the unbalanced transaction to be submitted, the old state and the new step, the step is run and Right the new state is returned. If the guard returns Just a, Left a is returned instead.

runStep

Arguments

:: forall w e state schema input. (AsSMContractError e, FromData state, ToData state, ToData input)

=> StateMachineClient state input
The state machine

-> input
The input to apply to the state machine

-> Contract w schema e (TransitionResult state input)

Run one step of a state machine, returning the new state.

runInitialise

Arguments

:: forall w e state schema input. (FromData state, ToData state, ToData input, AsSMContractError e)

=> StateMachineClient state input
The state machine

-> state
The initial state

-> Value
The value locked by the contract at the beginning

-> Contract w schema e state

Initialise a state machine

runGuardedStepWith

Arguments

:: forall w a e state schema input. (AsSMContractError e, FromData state, ToData state, ToData input)

=> ScriptLookups (StateMachine state input)
Additional lookups

-> TxConstraints input state
Additional constraints

-> StateMachineClient state input
The state machine

-> input
The input to apply to the state machine

-> (UnbalancedTx -> state -> state -> Maybe a)
The guard to check before running the step

-> Contract w schema e (Either a (TransitionResult state input))

The same as runGuardedStep but we can supply additional constraints and lookups for transaction.

runStepWith

Arguments

:: forall w e state schema input. (AsSMContractError e, FromData state, ToData state, ToData input)

=> ScriptLookups (StateMachine state input)
Additional lookups

-> TxConstraints input state
Additional constraints

-> StateMachineClient state input
The state machine

-> input
The input to apply to the state machine

-> Contract w schema e (TransitionResult state input)

Run one step of a state machine, returning the new state. We can supply additional constraints and lookups for transaction.

runInitialiseWith

Arguments

:: forall w e state schema input. (FromData state, ToData state, ToData input, AsSMContractError e)

=> ScriptLookups (StateMachine state input)
Additional lookups

-> TxConstraints input state
Additional constraints

-> StateMachineClient state input
The state machine

-> state
The initial state

-> Value
The value locked by the contract at the beginning

-> Contract w schema e state

Initialise a state machine and supply additional constraints and lookups for transaction.

getThreadToken :: AsSMContractError e => Contract w schema e ThreadToken

Create a thread token. The thread token contains a reference to an unspent output of the wallet, so it needs to used with mkStateMachine immediately, and the machine must be initialised, to prevent the output from getting spent in the mean time.

getOnChainState :: (AsSMContractError e, FromData state, ToData state) => StateMachineClient state i -> Contract w schema e (Maybe (OnChainState state i, Map TxOutRef DecoratedTxOut))

Get the current on-chain state of the state machine instance. Return Nothing if there is no state on chain. Throws an SMContractError if the number of outputs at the machine address is greater than one.

getStateData :: OnChainState s i -> s

waitForUpdate :: forall state i w schema e. (AsSMContractError e, AsContractError e, FromData state, ToData state, FromData i) => StateMachineClient state i -> Contract w schema e (Maybe (OnChainState state i))

Wait until the on-chain state of the state machine instance has changed, and return the new state, or return Nothing if the instance has been terminated. If waitForUpdate is called before the instance has even started then it returns the first state of the instance as soon as it has started.

waitForUpdateUntilSlot :: (AsSMContractError e, AsContractError e, FromData state, ToData state, FromData i) => StateMachineClient state i -> Slot -> Contract w schema e (WaitingResult Slot i state)

Wait for the on-chain state of the state machine instance to change until timeoutSlot, and return the new state, or return ContractEnded if the instance has been terminated. If waitForUpdate is called before the instance has even started then it returns the first state of the instance as soon as it has started.

waitForUpdateUntilTime :: (AsSMContractError e, AsContractError e, FromData state, ToData state, FromData i) => StateMachineClient state i -> POSIXTime -> Contract w schema e (WaitingResult POSIXTime i state)

Same as waitForUpdateUntilSlot, but works with POSIXTime instead.

waitForUpdateTimeout

Arguments

:: forall state i t w schema e. (AsSMContractError e, AsContractError e, FromData state, ToData state, FromData i)

=> StateMachineClient state i
The state machine client

-> Promise w schema e t
The timeout

-> Contract w schema e (Promise w schema e (WaitingResult t i (OnChainState state i)))

Construct a Promise that waits for an update to the state machine's on-chain state, or a user-defined timeout (whichever happens first).

Lower-level API

data StateMachineTransition state input

Constraints & lookups needed to transition a state machine instance

Constructors

StateMachineTransition

Fields

smtConstraints :: TxConstraints input state

smtOldState :: State state

smtNewState :: State state

smtLookups :: ScriptLookups (StateMachine state input)

mkStep :: forall w e state schema input. (AsSMContractError e, FromData state, ToData state) => StateMachineClient state input -> input -> Contract w schema e (Either (InvalidTransition state input) (StateMachineTransition state input))

Given a state machine client and an input to apply to the client's state machine instance, compute the StateMachineTransition that can produce an actual transaction performing the transition

Re-exports

data Void

Uninhabited data type

Since: base-4.8.0.0