Plutus.Contract.Trace

Description

A trace is a sequence of actions by simulated wallets that can be run on the mockchain. This module contains the functions needed to build traces.

Documentation

data TraceError e

Error produced while running a trace. Either a contract-specific error (of type e), or an AssertionError from the emulator.

Constructors

TraceAssertionError AssertionError

TContractError e

HookError EndpointError

data EndpointError

Constructors

EndpointNotActive (Maybe Wallet) EndpointDescription

class AsTraceError r e | r -> e where

Minimal complete definition

_TraceError

Methods

_TraceError :: Prism' r (TraceError e)

_TraceAssertionError :: Prism' r AssertionError

_TContractError :: Prism' r e

_HookError :: Prism' r EndpointError

toNotifyError :: ContractInstanceId -> EndpointError -> NotificationError

Handle contract requests

handleAdjustUnbalancedTx :: (Member (LogObserve (LogMessage Text)) effs, Member (LogMsg RequestHandlerLogMsg) effs, Member NodeClientEffect effs) => RequestHandler effs PABReq PABResp

handleSlotNotifications :: (Member (LogObserve (LogMessage Text)) effs, Member (LogMsg RequestHandlerLogMsg) effs, Member NodeClientEffect effs) => RequestHandler effs PABReq PABResp

handleTimeNotifications :: (Member (LogObserve (LogMessage Text)) effs, Member (LogMsg RequestHandlerLogMsg) effs, Member NodeClientEffect effs) => RequestHandler effs PABReq PABResp

handleOwnAddressesQueries :: (Member (LogObserve (LogMessage Text)) effs, Member WalletEffect effs) => RequestHandler effs PABReq PABResp

handleCurrentNodeClientSlotQueries :: (Member (LogObserve (LogMessage Text)) effs, Member NodeClientEffect effs) => RequestHandler effs PABReq PABResp

handleCurrentChainIndexSlotQueries :: (Member (LogObserve (LogMessage Text)) effs, Member ChainIndexQueryEffect effs) => RequestHandler effs PABReq PABResp

handleCurrentTimeQueries :: (Member (LogObserve (LogMessage Text)) effs, Member NodeClientEffect effs) => RequestHandler effs PABReq PABResp

handleCurrentNodeClientTimeRangeQueries :: (Member (LogObserve (LogMessage Text)) effs, Member NodeClientEffect effs) => RequestHandler effs PABReq PABResp

handleTimeToSlotConversions :: (Member (LogObserve (LogMessage Text)) effs, Member NodeClientEffect effs) => RequestHandler effs PABReq PABResp

handleUnbalancedTransactions :: (Member (LogObserve (LogMessage Text)) effs, Member (LogMsg RequestHandlerLogMsg) effs, Member WalletEffect effs) => RequestHandler effs PABReq PABResp

handlePendingTransactions :: (Member (LogObserve (LogMessage Text)) effs, Member (LogMsg RequestHandlerLogMsg) effs, Member WalletEffect effs) => RequestHandler effs PABReq PABResp

Submit the wallet's pending transactions to the blockchain.

handleChainIndexQueries :: (Member (LogObserve (LogMessage Text)) effs, Member ChainIndexQueryEffect effs) => RequestHandler effs PABReq PABResp

handleOwnInstanceIdQueries :: (Member (LogObserve (LogMessage Text)) effs, Member (Reader ContractInstanceId) effs) => RequestHandler effs PABReq PABResp

handleYieldedUnbalancedTx :: (Member (LogObserve (LogMessage Text)) effs, Member WalletEffect effs) => RequestHandler effs PABReq PABResp

handleGetParams :: (Member (LogObserve (LogMessage Text)) effs, Member NodeClientEffect effs) => RequestHandler effs PABReq PABResp

Initial distributions of emulated chains

type InitialDistribution = Map Wallet Value

defaultDist :: InitialDistribution

defaultDistFor :: [Wallet] -> InitialDistribution

Wallets

data Wallet

A wallet identifier

Constructors

Wallet

Fields

prettyWalletName :: Maybe String

getWalletId :: WalletId

mockWalletAddress :: Wallet -> CardanoAddress

Get the cardano address of a mock wallet. (Fails if the wallet is not a mock wallet).

mockWalletPaymentPubKey :: Wallet -> PaymentPubKey

The public key of a mock wallet. (Fails if the wallet is not a mock wallet).

mockWalletPaymentPubKeyHash :: Wallet -> PaymentPubKeyHash

The payment public key hash of a mock wallet. (Fails if the wallet is not a mock wallet).

knownWallets :: [Wallet]

knownWallet :: Integer -> Wallet