Plutus.Contract.Types

Documentation

type ContractEffs w e = '[Error e, LogMsg Value, Writer w, Checkpoint, Resumable PABResp PABReq]

Effects that are available to contracts.

handleContractEffs :: forall w e effs a. (Member (Error e) effs, Member (State CheckpointStore) effs, Member (State CheckpointKey) effs, Member (State (AccumState w)) effs, Member (LogMsg CheckpointLogMsg) effs, Member (LogMsg Value) effs, Monoid w) => Eff (ContractEffs w e) a -> Eff effs (Maybe (MultiRequestContStatus PABResp PABReq effs a))

newtype Contract w (s :: Row *) e a

Contract w s e a is a contract with schema s, producing a value of type a or an error e. See note [Contract Schema].

Constructors

Contract

Fields

unContract :: Eff (ContractEffs w e) a

class IsContract c where

Class of types that can be trivially converted to a Contract. For use with functions where it is convenient to accept both Contract and Promise types.

Methods

toContract :: c w s e a -> Contract w s e a

Select

newtype Promise w (s :: Row *) e a

A wrapper indicating that this contract starts with a waiting action. For use with select.

Constructors

Promise

Fields

awaitPromise :: Contract w s e a

promiseBind :: Promise w s e a -> (a -> Contract w s e b) -> Promise w s e b

Run more Contract code after the Promise.

promiseMap :: (Contract w1 s1 e1 a1 -> Contract w2 s2 e2 a2) -> Promise w1 s1 e1 a1 -> Promise w2 s2 e2 a2

Lift a mapping function for Contract to a mapping function for Promise.

select :: forall w s e a. Promise w s e a -> Promise w s e a -> Promise w s e a

select returns the contract that makes progress first, discarding the other one.

However, note that if multiples promises are chained together like P1 select P2 select P3 and all three can make progress at the same moment, then select will prioritize the promises starting from the right (first P3 then P2 then P1).

selectEither :: forall w s e a b. Promise w s e a -> Promise w s e b -> Promise w s e (Either a b)

A variant of select for contracts with different return types.

selectList :: [Promise w s e a] -> Contract w s e a

selectList returns the contract that makes progress first, discarding the other ones.

However, if multiple contracts can make progress, selectList prioritizes the ones appearing first in the input list. Therefore, the order of the list of promises is important.

never :: Promise w s e a

A Promise that is never fulfilled. This is the identity of select.

Error handling

data ContractError

Constructors

WalletContractError WalletAPIError

ChainIndexContractError Text ChainIndexResponse

EmulatorAssertionContractError AssertionError

ConstraintResolutionContractError MkTxError

TxConstraintResolutionContractError MkTxError

TxToCardanoConvertContractError ToCardanoError

ResumableContractError MatchingError

CCheckpointContractError CheckpointError

EndpointDecodeContractError

Fields

eeEndpointDescription :: EndpointDescription

The endpoint description which the decoding error occurred from

eeEndpointValue :: EndpointValue Value

The endpoint value that was used as an endpoint parameter

eeErrorMessage :: Text

JSON decoding error message

OtherContractError Text

class AsContractError r where

Minimal complete definition

_ContractError

Methods

_ContractError :: Prism' r ContractError

_WalletContractError :: Prism' r WalletAPIError

_ChainIndexContractError :: Prism' r (Text, ChainIndexResponse)

_EmulatorAssertionContractError :: Prism' r AssertionError

_ConstraintResolutionContractError :: Prism' r MkTxError

_TxConstraintResolutionContractError :: Prism' r MkTxError

_TxToCardanoConvertContractError :: Prism' r ToCardanoError

_ResumableContractError :: Prism' r MatchingError

_CCheckpointContractError :: Prism' r CheckpointError

_EndpointDecodeContractError :: Prism' r (EndpointDescription, EndpointValue Value, Text)

_OtherContractError :: Prism' r Text

newtype MatchingError

An error

Constructors

WrongVariantError

Fields

unWrongVariantError :: Text

mapError :: forall w s e e' a. (e -> e') -> Contract w s e a -> Contract w s e' a

Transform any exceptions thrown by the Contract using the given function.

throwError :: MonadError e m => e -> m a

Is used within a monadic computation to begin exception processing.

runError :: forall w s e e0 a. Contract w s e a -> Contract w s e0 (Either e a)

Turn a contract with error type e and return type a into one with any error type (ie. throwing no errors) that returns 'Either e a'

handleError :: forall w s e e' a. (e -> Contract w s e' a) -> Contract w s e a -> Contract w s e' a

Handle errors, potentially throwing new errors.

Checkpoints

class AsCheckpointError r where

Methods

_CheckpointError :: Prism' r CheckpointError

data CheckpointError

Constructors

JSONDecodeError Text

checkpoint :: forall w s e a. (AsCheckpointError e, FromJSON a, ToJSON a) => Contract w s e a -> Contract w s e a

Write the current state of the contract to a checkpoint.

checkpointLoop :: forall w s e a b. (AsCheckpointError e, FromJSON a, ToJSON a, ToJSON b, FromJSON b) => (a -> Contract w s e (Either b a)) -> a -> Contract w s e b

Run and update

runResumable :: Monoid w => [Response PABResp] -> CheckpointStore -> Eff (ContractEffs w e) a -> ResumableResult w e PABResp PABReq a

insertAndUpdate

Arguments

:: forall w e a. Monoid w

=> Eff (ContractEffs w e) a

-> CheckpointStore
Checkpoint store

-> Responses (CheckpointKey, PABResp)
Previous responses

-> Response PABResp

-> ResumableResult w e PABResp PABReq a

runWithRecord :: forall w e a. Monoid w => Eff (ContractEffs w e) a -> CheckpointStore -> Responses PABResp -> ResumableResult w e PABResp PABReq a

State

data ResumableResult w e i o a

The result of running a Resumable

Constructors

ResumableResult

Fields

_responses :: Responses (CheckpointKey, i)

_requests :: Requests o

_finalState :: Either e (Maybe a)

_logs :: Seq (LogMessage Value)

_lastLogs :: Seq (LogMessage Value)

_checkpointStore :: CheckpointStore

_observableState :: w

Accumulated, observable state of the contract

_lastState :: w

Last accumulated state

responses :: forall w e i o a i. Lens (ResumableResult w e i o a) (ResumableResult w e i o a) (Responses (CheckpointKey, i)) (Responses (CheckpointKey, i))

requests :: forall w e i o a o. Lens (ResumableResult w e i o a) (ResumableResult w e i o a) (Requests o) (Requests o)

finalState :: forall w e i o a e a. Lens (ResumableResult w e i o a) (ResumableResult w e i o a) (Either e (Maybe a)) (Either e (Maybe a))

logs :: forall w e i o a. Lens' (ResumableResult w e i o a) (Seq (LogMessage Value))

lastState :: forall w e i o a. Lens' (ResumableResult w e i o a) w

checkpointStore :: forall w e i o a. Lens' (ResumableResult w e i o a) CheckpointStore

observableState :: forall w e i o a. Lens' (ResumableResult w e i o a) w

shrinkResumableResult :: ResumableResult w e i o a -> ResumableResult w e i o a

Shrink the ResumableResult by deleting everything that's not needed to restore the state of the contract instance.

Run with continuations

data SuspendedContract w e i o a

Constructors

SuspendedContract

Fields

_resumableResult :: ResumableResult w e i o a

_continuations :: Maybe (MultiRequestContStatus i o (SuspendedContractEffects w e) a)

_checkpointKey :: CheckpointKey

resumableResult :: forall w e i o a. Lens' (SuspendedContract w e i o a) (ResumableResult w e i o a)

continuations :: forall w e i o a. Lens' (SuspendedContract w e i o a) (Maybe (MultiRequestContStatus i o (SuspendedContractEffects w e) a))

checkpointKey :: forall w e i o a. Lens' (SuspendedContract w e i o a) CheckpointKey

suspend

Arguments

:: forall w e a. Monoid w

=> CheckpointStore

-> Eff (ContractEffs w e) a
The contract

-> SuspendedContract w e PABResp PABReq a

Run an action of ContractEffs until it requests input for the first time, returning the SuspendedContract

runStep :: forall w e a. Monoid w => SuspendedContract w e PABResp PABReq a -> Response PABResp -> Maybe (SuspendedContract w e PABResp PABReq a)

Feed a Response to a SuspendedContract.

lastLogs :: forall w e i o a. Lens' (ResumableResult w e i o a) (Seq (LogMessage Value))