Plutus.Trace.Emulator.ContractInstance

Documentation

contractThread :: forall w s e effs. (Member (State EmulatorThreads) effs, Member (Error EmulatorRuntimeError) effs, ContractConstraints s, Show e, ToJSON e, ToJSON w, Monoid w) => ContractHandle w s e -> Eff (EmulatorAgentThreadEffs effs) ()

Start a new thread for a contract. Most of the work happens in runInstance.

getThread :: forall effs. (Member (State EmulatorThreads) effs, Member (Error EmulatorRuntimeError) effs) => ContractInstanceId -> Eff effs ThreadId

data EmulatorRuntimeError

runInstance :: forall w s e effs. (ContractConstraints s, Member (Error EmulatorRuntimeError) effs, Show e, ToJSON e, ToJSON w, Monoid w) => NetworkId -> Contract w s e () -> Maybe EmulatorMessage -> Eff (ContractInstanceThreadEffs w s e effs) ()

Run an instance of a contract

Instance state

data ContractInstanceState w (s :: Row *) e a

The state of a running contract instance with schema s and error type e Serialisable to JSON.

Constructors

ContractInstanceState

Fields

instContractState :: ResumableResult w e PABResp PABReq a

instEvents :: Seq (Response PABResp)

Events received by the contract instance. (Used for debugging purposes)

instHandlersHistory :: Seq [Request PABReq]

Requests issued by the contract instance (Used for debugging purposes)

emptyInstanceState :: forall w (s :: Row *) e a. Monoid w => Contract w s e a -> ContractInstanceStateInternal w s e a

addEventInstanceState :: forall w s e a. Monoid w => Response PABResp -> ContractInstanceStateInternal w s e a -> Maybe (ContractInstanceStateInternal w s e a)

Indexed block

data IndexedBlock

A block of transactions, indexed by tx outputs spent and by addresses on which new outputs are produced

Constructors

IndexedBlock

Fields

ibUtxoSpent :: Map TxOutRef ChainIndexTx

ibUtxoProduced :: Map CardanoAddress (NonEmpty ChainIndexTx)

indexBlock :: [ChainIndexTx] -> IndexedBlock

Internals

getHooks :: forall w s e effs. Member (State (ContractInstanceStateInternal w s e ())) effs => Eff effs [Request PABReq]

addResponse :: forall w s e effs. (Member (State (ContractInstanceStateInternal w s e ())) effs, Member (LogMsg ContractInstanceMsg) effs, Monoid w) => Response PABResp -> Eff effs ()

Add a Response to the contract instance state