Plutus.Trace.Emulator.Types

Documentation

data EmulatorMessage

Messages sent to, and received by, threads in the emulator.

Constructors

NewSlot [Block] Slot
A new slot has begun and some blocks were added.

EndpointCall ThreadId EndpointDescription Value
Call to an endpoint

Freeze
Tell the contract instance to freeze itself (see note [Freeze and Thaw])

ContractInstanceStateRequest ThreadId
Request for the current state of a contract instance

ContractInstanceStateResponse Value
Response to a contract instance state request

newtype EmulatorThreads

A map of contract instance ID to thread ID

Constructors

EmulatorThreads

Fields

_instanceIdThreads :: Map ContractInstanceId ThreadId

instanceIdThreads :: Iso' EmulatorThreads (Map ContractInstanceId ThreadId)

type EmulatorAgentThreadEffs effs = LogMsg ContractInstanceLog ': EmulatedWalletEffects' (Yield (AgentSystemCall EmulatorMessage) (Maybe EmulatorMessage) ': (Reader ThreadId ': effs))

Effects available to emulator agent threads. Includes emulated wallet effects and effects related to threading / waiting for messages.

type EmulatedWalletEffects = EmulatedWalletEffects' '[]

type EmulatedWalletEffects' effs = WalletEffect ': (Error WalletAPIError ': (NodeClientEffect ': (ChainIndexQueryEffect ': (LogObserve (LogMessage Text) ': (LogMsg RequestHandlerLogMsg ': (LogMsg TxBalanceMsg ': (LogMsg Text ': effs)))))))

Effects that are used to handle requests by contract instances. In the emulator these effects are handled by MultiAgent. In the PAB they are handled by the actual walletnodechain index, mediated by the PAB runtime.

data ContractInstanceTag

A user-defined tag for a contract instance. Used to find the instance's log messages in the emulator log.

Constructors

ContractInstanceTag

Fields

unContractInstanceTag :: Text

shortContractInstanceTag :: Text

walletInstanceTag :: Wallet -> ContractInstanceTag

The ContractInstanceTag for the contract instance of a wallet. See note [Wallet contract instances]

data ContractHandle w s e

A reference to a running contract in the emulator.

Constructors

ContractHandle

Fields

chContract :: Contract w s e ()

chInstanceId :: ContractInstanceId

chInstanceTag :: ContractInstanceTag

chNetworkId :: NetworkId

data Emulator

type ContractConstraints s = (Forall (Output s) Unconstrained1, Forall (Input s) Unconstrained1, AllUniqueLabels (Input s), AllUniqueLabels (Output s), Forall (Input s) FromJSON, Forall (Input s) ToJSON, Forall (Output s) FromJSON, Forall (Output s) ToJSON)

Instance state

data ContractInstanceState w (s :: Row *) e a

The state of a running contract instance with schema s and error type e Serialisable to JSON.

Constructors

ContractInstanceState

Fields

instContractState :: ResumableResult w e PABResp PABReq a

instEvents :: Seq (Response PABResp)

Events received by the contract instance. (Used for debugging purposes)

instHandlersHistory :: Seq [Request PABReq]

Requests issued by the contract instance (Used for debugging purposes)

data ContractInstanceStateInternal w (s :: Row *) e a

State of the contract instance, internal to the contract instance thread. It contains both the serialisable state of the contract instance and the non-serialisable continuations in SuspendedContract.

Constructors

ContractInstanceStateInternal

Fields

cisiSuspState :: SuspendedContract w e PABResp PABReq a

cisiEvents :: Seq (Response PABResp)

cisiHandlersHistory :: Seq [Request PABReq]

emptyInstanceState :: forall w (s :: Row *) e a. Monoid w => Contract w s e a -> ContractInstanceStateInternal w s e a

addEventInstanceState :: forall w s e a. Monoid w => Response PABResp -> ContractInstanceStateInternal w s e a -> Maybe (ContractInstanceStateInternal w s e a)

toInstanceState :: ContractInstanceStateInternal w (s :: Row *) e a -> ContractInstanceState w s e a

Extract the serialisable ContractInstanceState from the ContractInstanceStateInternal. We need to do this when we want to send the instance state to another thread.

Logging

data ContractInstanceLog

Constructors

ContractInstanceLog

Fields

_cilMessage :: ContractInstanceMsg

_cilId :: ContractInstanceId

_cilTag :: ContractInstanceTag

cilId :: Lens' ContractInstanceLog ContractInstanceId

cilMessage :: Lens' ContractInstanceLog ContractInstanceMsg

cilTag :: Lens' ContractInstanceLog ContractInstanceTag

data EmulatorRuntimeError

Constructors

ThreadIdNotFound ContractInstanceId

InstanceIdNotFound Wallet

EmulatorJSONDecodingError String Value

GenericError String

EmulatedWalletError WalletAPIError

AssertionError String

data ContractInstanceMsg

Log messages produced by contract instances

Constructors

Started

StoppedNoError

StoppedWithError String

ReceiveEndpointCall EndpointDescription Value

ReceiveEndpointCallSuccess

ReceiveEndpointCallFailure NotificationError

NoRequestsHandled

HandledRequest (Response Value)

CurrentRequests [Request Value]

InstErr EmulatorRuntimeError

ContractLog Value

SendingNotification Notification

NotificationSuccess Notification

NotificationFailure NotificationError

SendingContractState ThreadId

Freezing

_Started :: Prism' ContractInstanceMsg ()

_StoppedNoError :: Prism' ContractInstanceMsg ()

_StoppedWithError :: Prism' ContractInstanceMsg String

_ReceiveEndpointCall :: Prism' ContractInstanceMsg (EndpointDescription, Value)

_NoRequestsHandled :: Prism' ContractInstanceMsg ()

_HandledRequest :: Prism' ContractInstanceMsg (Response Value)

_CurrentRequests :: Prism' ContractInstanceMsg [Request Value]

_InstErr :: Prism' ContractInstanceMsg EmulatorRuntimeError

_ContractLog :: Prism' ContractInstanceMsg Value

data UserThreadMsg

Log message produced by the user (main) thread

Constructors

UserThreadErr EmulatorRuntimeError

UserLog String