Plutus.Trace.Scheduler

Documentation

data ThreadId

Unique identifier of a thread.

type SysCall effs systemEvent a = Either (MessageCall systemEvent) (ThreadCall effs systemEvent a)

data MessageCall systemEvent

Sending messages to other threads and waiting for new messages to arrive.

Constructors

WaitForMessage
Suspend ourselves (the caller) until we receive a message

Broadcast systemEvent
Send a message to all threads

Message ThreadId systemEvent
Send a message to a specific thread

data ThreadCall effs systemEvent a

The system calls we can make to the scheduler, affecting the the threads that are currently running.

Constructors

Fork (ThreadId -> SuspendedThread effs systemEvent a)
Start a new thread with a new thread ID.

Thaw ThreadId
Unfreeze a thread.

Exit a
Terminate the scheduler.

data WithPriority t

A thread with a Priority.

Constructors

WithPriority

Fields

_priority :: Priority

_thread :: t

data Priority

Priority of a thread.

Constructors

Normal
Thread is ready to run

Sleeping
Thread is sleeping, to be resumed only after an external event happens

Frozen
Thread is frozen, it will only be resumed after it is manually unfrozen via the Thaw sys call.

data Tag

A human-readable piece of data, used to identify threads and contract instances. See note [Thread Tag]

type EmSystemCall effs systemEvent a = WithPriority (SysCall effs systemEvent a)

type AgentSystemCall systemEvent = WithPriority (MessageCall systemEvent)

type SuspendedThread effs systemEvent a = WithPriority (EmThread effs systemEvent a)

data EmThread effs systemEvent a

Thread that can be run by the scheduler

Constructors

EmThread

Fields

_continuation :: Maybe systemEvent -> Eff effs (Status effs (EmSystemCall effs systemEvent a) (Maybe systemEvent) ())

The continuation to be run when the thread is resumed.

_threadId :: ThreadId

Thread ID

_tag :: Tag

Tag of the thread. See note [Thread Tag]

data SchedulerState effs systemEvent a

Scheduler state

Constructors

SchedulerState

Fields

_normalPrio :: Seq (EmThread effs systemEvent a)

Threads running at normal priority

_sleeping :: Seq (EmThread effs systemEvent a)

Sleeping threads (waiting for an external event)

_frozen :: Seq (EmThread effs systemEvent a)

Frozen threads (will not be resumed until they are explicitly unfrozen)

_lastThreadId :: ThreadId

Last thread id assigned to a thread

_mailboxes :: HashMap ThreadId (Seq systemEvent)

The mailboxes of all active threads.

_activeThreads :: Map Tag (HashSet ThreadId)

Map of tags to thread IDs. See note [Thread Tag]

Thread API

runThreads :: forall a effs systemEvent. (Eq systemEvent, Member (LogMsg SchedulerLog) effs) => Eff (Reader ThreadId ': (Yield (EmSystemCall effs systemEvent a) (Maybe systemEvent) ': effs)) () -> Eff effs (Maybe a)

Handle the 'Yield (EmSystemCall effs systemEvent) (Maybe systemEvent)' effect using the scheduler, see note [Scheduler]. runThreads only returns when all threads are finished, returning Nothing, or when exit is called, in which case the value passed to exit is returned.

fork

Arguments

:: forall effs systemEvent effs2 a. Member (Yield (EmSystemCall effs systemEvent a) (Maybe systemEvent)) effs2

=> Tag
Tag of the new thread. See note [Thread Tag]

-> Priority
Priority of the new thread.

-> Eff (Reader ThreadId ': (Yield (EmSystemCall effs systemEvent a) (Maybe systemEvent) ': effs)) ()

-> Eff effs2 (Maybe systemEvent)

Start a new thread

sleep :: forall effs systemEvent effs2 a. Member (Yield (EmSystemCall effs systemEvent a) (Maybe systemEvent)) effs2 => Priority -> Eff effs2 (Maybe systemEvent)

Suspend the current thread

exit :: forall effs systemEvent effs2 a. Member (Yield (EmSystemCall effs systemEvent a) (Maybe systemEvent)) effs2 => a -> Eff effs2 (Maybe systemEvent)

Stop the scheduler and let it return with the given value.

Etc.

mkThread :: Tag -> Priority -> Eff (Reader ThreadId ': (Yield (EmSystemCall effs systemEvent a) (Maybe systemEvent) ': effs)) () -> ThreadId -> SuspendedThread effs systemEvent a

Make a thread with the given priority from an action. This is a convenience for defining SimulatorInterpreter values.

mkSysCall

Arguments

:: forall effs systemEvent effs2 a. Member (Yield (EmSystemCall effs systemEvent a) (Maybe systemEvent)) effs2

=> Priority
The Priority of the caller

-> SysCall effs systemEvent a
The system call

-> Eff effs2 (Maybe systemEvent)

Make a system call

mkAgentSysCall

Arguments

:: forall effs systemEvent. Member (Yield (AgentSystemCall systemEvent) (Maybe systemEvent)) effs

=> Priority
The Priority of the caller

-> MessageCall systemEvent
The system call

-> Eff effs (Maybe systemEvent)

Make a MessageCall system call for some agent

data SchedulerLog

Constructors

SchedulerLog

Fields

slEvent :: ThreadEvent

slThread :: ThreadId

slTag :: Tag

slPrio :: Priority

data ThreadEvent

Constructors

Stopped

Resumed

Suspended

Started

Thawed