Wallet.API

Description

Mock wallet implementation

Documentation

data WalletEffect r

submitTxn :: forall effs. Member WalletEffect effs => CardanoTx -> Eff effs ()

ownPaymentPubKeyHash :: (Member WalletEffect effs, Member (Error WalletAPIError) effs) => Eff effs PaymentPubKeyHash

Deprecated: Use ownFirstPaymentPubKeyHash, ownPaymentPubKeyHashes or ownAddresses instead

ownPaymentPubKeyHashes :: Member WalletEffect effs => Eff effs [PaymentPubKeyHash]

ownFirstPaymentPubKeyHash :: (Member WalletEffect effs, Member (Error WalletAPIError) effs) => Eff effs PaymentPubKeyHash

ownAddresses :: forall effs. Member WalletEffect effs => Eff effs (NonEmpty CardanoAddress)

balanceTx :: forall effs. Member WalletEffect effs => UnbalancedTx -> Eff effs (Either WalletAPIError CardanoTx)

yieldUnbalancedTx :: forall effs. Member WalletEffect effs => UnbalancedTx -> Eff effs ()

data NodeClientEffect r

publishTx :: forall effs. Member NodeClientEffect effs => CardanoTx -> Eff effs ()

getClientSlot :: forall effs. Member NodeClientEffect effs => Eff effs Slot

getClientParams :: forall effs. Member NodeClientEffect effs => Eff effs Params

newtype PubKey

Constructors

PubKey

Fields

getPubKey :: LedgerBytes

newtype PubKeyHash

The hash of a public key. This is frequently used to identify the public key, rather than the key itself.

Constructors

PubKeyHash

Fields

getPubKeyHash :: BuiltinByteString

signTxAndSubmit :: Member WalletEffect effs => CardanoTx -> Eff effs CardanoTx

Add the wallet's signature to the transaction and submit it. Returns the transaction with the wallet's signature.

signTxAndSubmit_ :: Member WalletEffect effs => CardanoTx -> Eff effs ()

A version of signTxAndSubmit that discards the result.

payToAddress :: (Member WalletEffect effs, Member (Error WalletAPIError) effs, Member (LogMsg Text) effs, Member (LogMsg RequestHandlerLogMsg) effs) => Params -> SlotRange -> Value -> Address -> Eff effs CardanoTx

Transfer some funds to an address, returning the transaction that was submitted.

Note: Due to a constraint in the Cardano ledger, each tx output must have a minimum amount of Ada. Therefore, the funds to transfer will be adjusted to satisfy that constraint. See adjustUnbalancedTx.

payToAddress_ :: (Member WalletEffect effs, Member (Error WalletAPIError) effs, Member (LogMsg Text) effs, Member (LogMsg RequestHandlerLogMsg) effs) => Params -> SlotRange -> Value -> Address -> Eff effs ()

Transfer some funds to an address.

payToPaymentPublicKeyHash :: (Member WalletEffect effs, Member (Error WalletAPIError) effs, Member (LogMsg Text) effs, Member (LogMsg RequestHandlerLogMsg) effs) => Params -> SlotRange -> Value -> PaymentPubKeyHash -> Eff effs CardanoTx

Transfer some funds to an address locked by a public key, returning the transaction that was submitted.

Note: Due to a constraint in the Cardano ledger, each tx output must have a minimum amount of Ada. Therefore, the funds to transfer will be adjusted to satisfy that constraint. See adjustUnbalancedTx.

payToPaymentPublicKeyHash_ :: (Member WalletEffect effs, Member (Error WalletAPIError) effs, Member (LogMsg Text) effs, Member (LogMsg RequestHandlerLogMsg) effs) => Params -> SlotRange -> Value -> PaymentPubKeyHash -> Eff effs ()

Transfer some funds to an address locked by a public key.

data Params

Constructors

Params

Fields

pSlotConfig :: SlotConfig

emulatorPParams :: PParams

Convert Params to cardano-ledger PParams

pNetworkId :: NetworkId

Slot ranges

data Interval a

An interval of as.

The interval may be either closed or open at either end, meaning that the endpoints may or may not be included in the interval.

The interval can also be unbounded on either side.

Constructors

Interval

Fields

ivFrom :: LowerBound a

ivTo :: UpperBound a

data Slot

The slot number. This is a good proxy for time, since on the Cardano blockchain slots pass at a constant rate.

type SlotRange = Interval Slot

An Interval of Slots.

width :: SlotRange -> Maybe Integer

Number of Slots covered by the interval, if finite. width (from x) == Nothing.

defaultSlotRange :: SlotRange

The default slot validity range for transactions.

interval :: a -> a -> Interval a

interval a b includes all values that are greater than or equal to a and smaller than or equal to b. Therefore it includes a and b.

singleton :: a -> Interval a

isEmpty :: (Enum a, Ord a) => Interval a -> Bool

Check if an Interval is empty.

always :: Interval a

An Interval that covers every slot.

member :: Ord a => a -> Interval a -> Bool

Check whether a value is in an interval.

before :: Ord a => a -> Interval a -> Bool

Check if a value is earlier than the beginning of an Interval.

after :: Ord a => a -> Interval a -> Bool

Check if a value is later than the end of a Interval.

contains :: Ord a => Interval a -> Interval a -> Bool

a contains b is true if the Interval b is entirely contained in a. That is, a contains b if for every entry s, if member s b then member s a.

Error handling

data WalletAPIError

An error thrown by wallet interactions.

Constructors

InsufficientFunds Text
There were insufficient funds to perform the desired operation.

ChangeHasLessThanNAda Value Ada
The change when selecting coins contains less than the minimum amount of Ada.

NoPaymentPubKeyHashError
The wallet doesn't have any payment key hash, which should not be possible.

PaymentPrivateKeyNotFound PaymentPubKeyHash
The private key of this public key hash is not known to the wallet.

ValidationError ValidationError
There was an error during off-chain validation.

ToCardanoError ToCardanoError
There was an error while converting to Cardano.API format.

PaymentMkTxError MkTxError
There was an error while creating a payment transaction

RemoteClientFunctionNotYetSupported Text
The called wallet effect is not yet supported in a remote wallet client scenario.

OtherError Text
Some other error occurred.

throwInsufficientFundsError :: Member (Error WalletAPIError) effs => Text -> Eff effs a

throwOtherError :: Member (Error WalletAPIError) effs => Text -> Eff effs a