Wallet.Emulator.Error

Documentation

data WalletAPIError

An error thrown by wallet interactions.

Constructors

InsufficientFunds Text
There were insufficient funds to perform the desired operation.

ChangeHasLessThanNAda Value Ada
The change when selecting coins contains less than the minimum amount of Ada.

NoPaymentPubKeyHashError
The wallet doesn't have any payment key hash, which should not be possible.

PaymentPrivateKeyNotFound PaymentPubKeyHash
The private key of this public key hash is not known to the wallet.

ValidationError ValidationError
There was an error during off-chain validation.

ToCardanoError ToCardanoError
There was an error while converting to Cardano.API format.

PaymentMkTxError MkTxError
There was an error while creating a payment transaction

RemoteClientFunctionNotYetSupported Text
The called wallet effect is not yet supported in a remote wallet client scenario.

OtherError Text
Some other error occurred.

throwInsufficientFundsError :: Member (Error WalletAPIError) effs => Text -> Eff effs a

throwOtherError :: Member (Error WalletAPIError) effs => Text -> Eff effs a