Wallet.Emulator.MultiAgent

Documentation

data Assertion

Assertions which will be checked during execution of the emulator.

Constructors

IsValidated CardanoTx
Assert that the given transaction is validated.

OwnFundsEqual Wallet Value
Assert that the funds belonging to a wallet's public-key address are equal to a value.

data EmulatorTimeEvent e

An event with a timestamp measured in emulator time (currently: Slot)

Constructors

EmulatorTimeEvent

Fields

_eteEmulatorTime :: Slot

_eteEvent :: e

eteEvent :: forall e e. Lens (EmulatorTimeEvent e) (EmulatorTimeEvent e) e e

eteEmulatorTime :: forall e. Lens' (EmulatorTimeEvent e) Slot

emulatorTimeEvent :: Slot -> Prism' (EmulatorTimeEvent e) e

data EmulatorEvent'

Events produced by the blockchain emulator.

Constructors

ChainEvent ChainEvent

ClientEvent Wallet NodeClientEvent

WalletEvent Wallet WalletEvent

ChainIndexEvent Wallet ChainIndexLog

SchedulerEvent SchedulerLog

InstanceEvent ContractInstanceLog

UserThreadEvent UserThreadMsg

type EmulatorEvent = EmulatorTimeEvent EmulatorEvent'

chainEvent :: Prism' EmulatorEvent' ChainEvent

walletClientEvent :: Wallet -> Prism' EmulatorEvent' NodeClientEvent

walletEvent :: Wallet -> Prism' EmulatorEvent' WalletEvent

walletEvent' :: Prism' EmulatorEvent' (Wallet, WalletEvent)

chainIndexEvent :: Wallet -> Prism' EmulatorEvent' ChainIndexLog

schedulerEvent :: Prism' EmulatorEvent' SchedulerLog

instanceEvent :: Prism' EmulatorEvent' ContractInstanceLog

userThreadEvent :: Prism' EmulatorEvent' UserThreadMsg

type EmulatedWalletControlEffects = '[NodeClientControlEffect, ChainIndexControlEffect, SigningProcessControlEffect, LogObserve (LogMessage Text), LogMsg Text]

data MultiAgentEffect r where

The type of actions in the emulator.

Constructors

WalletAction :: Wallet -> Eff EmulatedWalletEffects r -> MultiAgentEffect r
A direct action performed by a wallet. Usually represents a "user action", as it is triggered externally.

data MultiAgentControlEffect r where

Constructors

WalletControlAction :: Wallet -> Eff EmulatedWalletControlEffects r -> MultiAgentControlEffect r
An action affecting the emulated parts of a wallet (only available in emulator - see note [Control effects].)

Assertion :: Assertion -> MultiAgentControlEffect ()
An assertion in the event stream, which can inspect the current state.

walletAction :: Member MultiAgentEffect effs => Wallet -> Eff EmulatedWalletEffects r -> Eff effs r

Run an action in the context of a wallet (ie. agent)

handleMultiAgentEffects :: forall effs. Member MultiAgentEffect effs => Wallet -> Eff (EmulatedWalletEffects' effs) ~> Eff effs

raiseWallet :: forall f effs. (Member f EmulatedWalletEffects, Member MultiAgentEffect effs) => Wallet -> f ~> Eff effs

walletControlAction :: Member MultiAgentControlEffect effs => Wallet -> Eff EmulatedWalletControlEffects r -> Eff effs r

Run a control action in the context of a wallet

assertion :: Member MultiAgentControlEffect effs => Assertion -> Eff effs ()

assertOwnFundsEq :: Member MultiAgentControlEffect effs => Wallet -> Value -> Eff effs ()

Issue an assertion that the funds for a given wallet have the given value.

assertIsValidated :: Member MultiAgentControlEffect effs => CardanoTx -> Eff effs ()

Issue an assertion that the given transaction has been validated.

data EmulatorState

The state of the emulator itself.

Constructors

EmulatorState

Fields

_chainState :: ChainState

Mockchain

_walletStates :: Map Wallet WalletState

The state of each agent.

_emulatorLog :: [LogMessage EmulatorEvent]

The emulator log messages, with the newest last.

walletStates :: Lens' EmulatorState (Map Wallet WalletState)

emulatorLog :: Lens' EmulatorState [LogMessage EmulatorEvent]

chainState :: Lens' EmulatorState ChainState

walletState :: Wallet -> Lens' EmulatorState WalletState

chainOldestFirst :: Lens' EmulatorState Blockchain

Get the blockchain as a list of blocks, starting with the oldest (genesis) block.

chainUtxo :: Getter EmulatorState AddressMap

fundsDistribution :: EmulatorState -> Map Wallet Value

Get a map with the total value of each wallet's "own funds".

emLog :: EmulatorState -> [LogMessage EmulatorEvent]

Get the emulator log.

emptyEmulatorState :: EmulatorState

emulatorState :: Blockchain -> EmulatorState

Initialise the emulator state with a blockchain.

emulatorStatePool :: TxPool -> EmulatorState

Initialise the emulator state with a pool of pending transactions.

emulatorStateInitialDist :: Params -> Map PaymentPubKeyHash Value -> Either ToCardanoError EmulatorState

Initialise the emulator state with a single pending transaction that creates the initial distribution of funds to public key addresses.

type MultiAgentEffs = '[State EmulatorState, LogMsg EmulatorEvent', Error WalletAPIError, Error ChainIndexError, Error AssertionError, ChainEffect, ChainControlEffect]

handleMultiAgentControl :: forall effs. Members MultiAgentEffs effs => Eff (MultiAgentControlEffect ': effs) ~> Eff effs

handleMultiAgent :: forall effs. Members MultiAgentEffs effs => Eff (MultiAgentEffect ': effs) ~> Eff effs

assert :: Members MultiAgentEffs effs => Assertion -> Eff effs ()

Issue an Assertion.

ownFundsEqual :: Members MultiAgentEffs effs => Wallet -> Value -> Eff effs ()

Issue an assertion that the funds for a given wallet have the given value.

isValidated :: Members MultiAgentEffs effs => CardanoTx -> Eff effs ()

Issue an assertion that the given transaction has been validated.

_singleton :: AReview [a] a