Wallet.Emulator.Types

Wallets

data Wallet

A wallet identifier

Constructors

Wallet

Fields

prettyWalletName :: Maybe String

getWalletId :: WalletId

newtype WalletId

Constructors

WalletId

Fields

unWalletId :: Digest Blake2b_160

data XPrv

data XPub

mockWalletAddress :: Wallet -> CardanoAddress

Get the cardano address of a mock wallet. (Fails if the wallet is not a mock wallet).

mockWalletPaymentPubKey :: Wallet -> PaymentPubKey

The public key of a mock wallet. (Fails if the wallet is not a mock wallet).

mockWalletPaymentPubKeyHash :: Wallet -> PaymentPubKeyHash

The payment public key hash of a mock wallet. (Fails if the wallet is not a mock wallet).

knownWallets :: [Wallet]

knownWallet :: Integer -> Wallet

newtype WalletNumber

Constructors

WalletNumber

Fields

getWallet :: Integer

toWalletNumber :: MockWallet -> WalletNumber

fromWalletNumber :: WalletNumber -> Wallet

data MockWallet

Constructors

MockWallet

Fields

mwWalletId :: Digest Blake2b_160

mwPaymentKey :: MockPrivateKey

mwStakeKey :: Maybe MockPrivateKey

mwPrintAs :: Maybe String

type TxPool = [CardanoTx]

Emulator

type EmulatorEffs = '[MultiAgentEffect, ChainEffect, ChainControlEffect]

newtype AssertionError

An error emitted when an Assertion fails.

Constructors

GenericAssertion

Fields

unAssertionError :: Text

class AsAssertionError r where

Minimal complete definition

_AssertionError

Methods

_AssertionError :: Prism' r AssertionError

_GenericAssertion :: Prism' r Text

data ChainClientNotification

Constructors

BlockValidated Block

SlotChanged Slot

type EmulatorEvent = EmulatorTimeEvent EmulatorEvent'

data EmulatorEvent'

Events produced by the blockchain emulator.

data EmulatorTimeEvent e

An event with a timestamp measured in emulator time (currently: Slot)

Constructors

EmulatorTimeEvent

Fields

_eteEmulatorTime :: Slot

_eteEvent :: e

Wallet state

data WalletState

The state used by the mock wallet environment.

Constructors

WalletState

Fields

_mockWallet :: MockWallet

Mock wallet with the user's private key.

_nodeClient :: NodeClientState

The representation of the node, as known by the wallet

_chainIndexEmulatorState :: ChainIndexEmulatorState

the chain index info known by the wallet

_signingProcess :: Maybe SigningProcess

Override the signing process. Used for testing multi-agent use cases.

emptyWalletState :: Wallet -> Maybe WalletState

Empty wallet state for an emulator Wallet. Returns Nothing if the wallet is not known in the emulator.

ownPaymentPrivateKey :: WalletState -> PaymentPrivateKey

ownAddress :: WalletState -> CardanoAddress

Get the user's own payment public-key address.

Traces

walletAction :: Member MultiAgentEffect effs => Wallet -> Eff EmulatedWalletEffects r -> Eff effs r

Run an action in the context of a wallet (ie. agent)

Emulator internals

data EmulatorState

The state of the emulator itself.

Constructors

EmulatorState

Fields

_chainState :: ChainState

Mockchain

_walletStates :: Map Wallet WalletState

The state of each agent.

_emulatorLog :: [LogMessage EmulatorEvent]

The emulator log messages, with the newest last.

emptyEmulatorState :: EmulatorState

emulatorState :: Blockchain -> EmulatorState

Initialise the emulator state with a blockchain.

emulatorStatePool :: TxPool -> EmulatorState

Initialise the emulator state with a pool of pending transactions.

emulatorStateInitialDist :: Params -> Map PaymentPubKeyHash Value -> Either ToCardanoError EmulatorState

Initialise the emulator state with a single pending transaction that creates the initial distribution of funds to public key addresses.

txPool :: Lens' ChainState TxPool

walletStates :: Lens' EmulatorState (Map Wallet WalletState)

index :: Lens' ChainState UtxoIndex

chainState :: Lens' EmulatorState ChainState

chainCurrentSlot :: Lens' ChainState Slot

processEmulated :: forall effs. (Member (Error WalletAPIError) effs, Member (Error ChainIndexError) effs, Member (State EmulatorState) effs, Member (LogMsg EmulatorEvent') effs) => Params -> Eff (MultiAgentEffect ': (MultiAgentControlEffect ': (ChainEffect ': (ChainControlEffect ': effs)))) ~> Eff effs

fundsDistribution :: EmulatorState -> Map Wallet Value

Get a map with the total value of each wallet's "own funds".

emLog :: EmulatorState -> [LogMessage EmulatorEvent]

Get the emulator log.