Wallet.Emulator.Wallet

Documentation

newtype SigningProcess

Constructors

SigningProcess

Fields

unSigningProcess :: forall effs. Member (Error WalletAPIError) effs => [PaymentPubKeyHash] -> CardanoTx -> Eff effs CardanoTx

data Wallet

A wallet identifier

Constructors

Wallet

Fields

prettyWalletName :: Maybe String

getWalletId :: WalletId

toMockWallet :: MockWallet -> Wallet

knownWallets :: [Wallet]

knownWallet :: Integer -> Wallet

fromWalletNumber :: WalletNumber -> Wallet

newtype WalletId

Constructors

WalletId

Fields

unWalletId :: WalletId

toBase16 :: WalletId -> Text

fromBase16 :: Text -> Either String WalletId

walletToMockWallet :: Wallet -> Maybe MockWallet

The MockWallet whose ID is the given wallet ID (if it exists)

walletToMockWallet' :: Wallet -> MockWallet

The same as walletToMockWallet but fails with an error instead of returning Nothing.

mockWalletPaymentPubKey :: Wallet -> PaymentPubKey

The public key of a mock wallet. (Fails if the wallet is not a mock wallet).

mockWalletPaymentPubKeyHash :: Wallet -> PaymentPubKeyHash

The payment public key hash of a mock wallet. (Fails if the wallet is not a mock wallet).

mockWalletAddress :: Wallet -> CardanoAddress

Get the cardano address of a mock wallet. (Fails if the wallet is not a mock wallet).

data WalletEvent

Constructors

GenericLog Text

CheckpointLog CheckpointLogMsg

RequestHandlerLog RequestHandlerLogMsg

TxBalanceLog TxBalanceMsg

_TxBalanceLog :: Prism' WalletEvent TxBalanceMsg

_RequestHandlerLog :: Prism' WalletEvent RequestHandlerLogMsg

_CheckpointLog :: Prism' WalletEvent CheckpointLogMsg

_GenericLog :: Prism' WalletEvent Text

data WalletState

The state used by the mock wallet environment.

Constructors

WalletState

Fields

_mockWallet :: MockWallet

Mock wallet with the user's private key.

_nodeClient :: NodeClientState

The representation of the node, as known by the wallet

_chainIndexEmulatorState :: ChainIndexEmulatorState

the chain index info known by the wallet

_signingProcess :: Maybe SigningProcess

Override the signing process. Used for testing multi-agent use cases.

signingProcess :: Lens' WalletState (Maybe SigningProcess)

nodeClient :: Lens' WalletState NodeClientState

mockWallet :: Lens' WalletState MockWallet

chainIndexEmulatorState :: Lens' WalletState ChainIndexEmulatorState

ownPaymentPrivateKey :: WalletState -> PaymentPrivateKey

ownPaymentPublicKey :: WalletState -> PaymentPubKey

ownAddress :: WalletState -> CardanoAddress

Get the user's own payment public-key address.

fromMockWallet :: MockWallet -> WalletState

An empty wallet using the given private key. for that wallet as the sole watched address.

emptyWalletState :: Wallet -> Maybe WalletState

Empty wallet state for an emulator Wallet. Returns Nothing if the wallet is not known in the emulator.

handleWallet :: (Member (Error WalletAPIError) effs, Member NodeClientEffect effs, Member ChainIndexQueryEffect effs, Member (State WalletState) effs, Member (LogMsg TxBalanceMsg) effs) => WalletEffect ~> Eff effs

handleBalance :: (Member NodeClientEffect effs, Member ChainIndexQueryEffect effs, Member (State WalletState) effs, Member (LogMsg TxBalanceMsg) effs, Member (Error WalletAPIError) effs) => UnbalancedTx -> Eff effs CardanoTx

handleAddSignature :: (Member (State WalletState) effs, Member (Error WalletAPIError) effs) => CardanoTx -> Eff effs CardanoTx

ownOutputs :: forall effs. Member ChainIndexQueryEffect effs => WalletState -> Eff effs (Map TxOutRef DecoratedTxOut)

defaultSigningProcess :: MockWallet -> SigningProcess

The default signing process is signWallet

signWithPrivateKey :: PaymentPrivateKey -> SigningProcess

signWallet :: MockWallet -> SigningProcess

Sign the transaction by calling signTxnWithKey (throwing a PrivateKeyNotFound error if called with a key other than the wallet's private key)

signTxnWithKey :: Member (Error WalletAPIError) r => MockWallet -> CardanoTx -> PaymentPubKeyHash -> Eff r CardanoTx

Sign the transaction with the private key of the mock wallet.

signTxWithPrivateKey :: Member (Error WalletAPIError) r => PaymentPrivateKey -> CardanoTx -> PaymentPubKeyHash -> Eff r CardanoTx

Sign the transaction with the private key, if the hash is that of the private key.

signPrivateKeys :: [PaymentPrivateKey] -> SigningProcess

Sign the transaction with the given private keys, ignoring the list of public keys that the SigningProcess is passed.

data SigningProcessControlEffect r where

Constructors

SetSigningProcess :: Maybe SigningProcess -> SigningProcessControlEffect ()

setSigningProcess :: forall effs. Member SigningProcessControlEffect effs => Maybe SigningProcess -> Eff effs ()

type SigningProcessEffs = '[State (Maybe SigningProcess), Error WalletAPIError]

handleSigningProcessControl :: Members SigningProcessEffs effs => Eff (SigningProcessControlEffect ': effs) ~> Eff effs

data Entity

An Entity is a thing that can hold Value. Used in the balances function to compute who holds for a given chain state and set of wallets.

Constructors

WalletEntity Wallet

PubKeyHashEntity PubKeyHash

ScriptEntity ValidatorHash

type WalletSet = Map Wallet WalletState

walletPaymentPubKeyHashes :: WalletSet -> Map PaymentPubKeyHash Wallet

Pick out all the public keys from the set of wallets and map them back to their corresponding wallets.

balances :: ChainState -> WalletSet -> Map Entity Value

For a set of wallets, convert them into a map of value: entity, where entity is one of Entity.

Orphan instances

Data WalletId

Instance details

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> WalletId -> c WalletId

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c WalletId

toConstr :: WalletId -> Constr

dataTypeOf :: WalletId -> DataType

dataCast1 :: Typeable t => (forall d. Data d => c (t d)) -> Maybe (c WalletId)

dataCast2 :: Typeable t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c WalletId)

gmapT :: (forall b. Data b => b -> b) -> WalletId -> WalletId

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> WalletId -> r

gmapQr :: forall r r'. (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> WalletId -> r

gmapQ :: (forall d. Data d => d -> u) -> WalletId -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> WalletId -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> WalletId -> m WalletId

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> WalletId -> m WalletId

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> WalletId -> m WalletId

ToSchema WalletId

Instance details

Methods

declareNamedSchema :: Proxy WalletId -> Declare (Definitions Schema) NamedSchema