Wallet.Graph

Description

Support for visualisation of a blockchain as a graph.

Documentation

txnFlows :: [PubKey] -> Blockchain -> [FlowLink]

Compute the FlowLinks for a Blockchain given a set of known PubKeys.

graph :: [FlowLink] -> FlowGraph

Construct a graph from a list of FlowLinks.

data FlowGraph

The flow graph, consisting of a set of nodes (TxRefs) and edges (FlowLinks).

data FlowLink

A link in the flow graph.

data TxRef

A wrapper around the first 8 digits of a TxId.

data UtxOwner

The owner of an unspent transaction output.

data UtxoLocation

The location of a transaction in a blockchain specified by two indices: the index of the containing block in the chain, and the index of the transaction within the block.